### 2025-09-15 d79d2e6
speedup if numba is missing, bump version
- numba not used if fortran subroutines are compiled
(numba is now only a fallback speedup method)

- changed a default in plotdisloc method (faster for accelerating screw case)

- remove some obsolete/unused code
