+++ pydislocdyn/__init__.py
@@ -2,7 +2,7 @@
 # PyDislocDyn
 # Author: Daniel N. Blaschke
 # Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-# Date: Jan. 8, 2024 - Sept. 12, 2025
+# Date: Jan. 8, 2024 - Sept. 14, 2025
 '''PyDislocDyn is a suite of python programs designed to perform various calculations for basic research
    in dislocation dynamics in metals with various crystal symmetries in the continuum limit. In particular,
    one of its main purposes is to calculate dislocation drag from phonon wind. Additional features include
@@ -17,7 +17,7 @@ from .crystals import IsoAverages, metal_props, Miller_to_Cart
 from .dislocations import StrohGeometry, Dislocation, readinputfile
 from .phononwind import elasticA3, phonondrag, B_of_sigma
 
-__version__ = '1.3.2+dev'
+__version__ = '1.3.3-beta'
 
 if usefortran:
     from .utilities import fsub as subroutines
