### 2025-09-15 340506c
move joss paper file to root folder
joss editorial bot had trouble finding it previously
