### 2025-09-15 b1df8c7
fix paper path in workflow
