+++ pydislocdyn/dislocations/steadystate.py
@@ -1,7 +1,7 @@
 # Compute various properties of a moving dislocation
 # Author: Daniel N. Blaschke
 # Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-# Date: Nov. 3, 2017 - Sept. 13, 2025
+# Date: Nov. 3, 2017 - Sept. 15, 2025
 '''This submodule contains a class, StrohGeometry, to calculate the displacement field of a steady state dislocation
    as well as various other properties. See also the more general Dislocation class defined in pydislocdyn.dislocations.general,
    which inherits from the StrohGeometry class defined here and the metal_props class defined in pydislocdyn.crystals.'''
@@ -50,7 +50,7 @@ class StrohGeometry:
         
         bsq = np.dot(self.b,self.b)
         nsq = np.dot(self.n0,self.n0)
-        if isinstance(bsq, sp.Expr) or isinstance(bsq, sp.Expr):
+        if isinstance(bsq, sp.Expr) or isinstance(nsq, sp.Expr):
             self.b = np.array(sp.simplify(self.b/sp.sqrt(bsq)))
             self.n0 = np.array(sp.simplify(self.n0/sp.sqrt(nsq)))
             self.t = np.empty(self.t.shape,dtype=object)
@@ -58,24 +58,25 @@ class StrohGeometry:
             for i,th in enumerate(self.theta*sp.pi/np.pi):
                 self.t[i] = sp.matrix2numpy(sp.simplify(sp.Matrix(self.b*sp.cos(th)) + sp.Matrix(self.b*sp.sin(th)).cross(sp.Matrix(self.n0)))).reshape((3))
                 self.m0[i] = sp.matrix2numpy(sp.simplify(sp.Matrix(self.n0).cross(sp.Matrix(self.t[i])))).reshape((3))
-            return -1 ## skip the rest: would need numbers, not sympy symbols
-        if bsq>1e-12 and abs(bsq-1)>1e-12:
-            self.b = self.b/np.sqrt(bsq)
-        if nsq>1e-12 and abs(nsq-1)>1e-12:
-            self.n0 = self.n0/np.sqrt(nsq)
-        if np.abs(self.b @ self.n0)>1e-12:
-            print(f"WARNING: the Burgers vector is not normal to the slip plane normal for Dislocation {self.name}: \n{self.b @ self.n0=:.6f} != 0")
-        
-        self.t = np.outer(np.cos(self.theta),self.b) + np.outer(np.sin(self.theta),np.cross(self.b,self.n0))
-        self.m0 = np.cross(self.n0,self.t)
-        
-        for i in range(3):
-            self.M[i] = np.outer(self.m0[:,i],np.cos(self.phi)) + np.outer(np.repeat(self.n0[i],Ntheta),np.sin(self.phi))
-            self.N[i] = np.outer(np.repeat(self.n0[i],Ntheta),np.cos(self.phi)) - np.outer(self.m0[:,i],np.sin(self.phi))
-            for j in range(3):
-                for k in range(3):
-                    for l in range(3):
-                        self.Cv[i,j,k,l] = self.m0[:,i]*delta[j,k]*self.m0[:,l]
+            ## skip the rest: would need numbers, not sympy symbols
+        else:
+            if bsq>1e-12 and abs(bsq-1)>1e-12:
+                self.b = self.b/np.sqrt(bsq)
+            if nsq>1e-12 and abs(nsq-1)>1e-12:
+                self.n0 = self.n0/np.sqrt(nsq)
+            if np.abs(self.b @ self.n0)>1e-12:
+                print(f"WARNING: the Burgers vector is not normal to the slip plane normal: \n{self.b @ self.n0=:.6f} != 0")
[…]
