+++ pydislocdyn/phononwind.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # Author: Daniel N. Blaschke
 # Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-# Date: Nov. 5, 2017 - Sept. 13, 2025
+# Date: Nov. 5, 2017 - Sept. 16, 2025
 '''This module implements the calculation of a dislocation drag coefficient from phonon wind.
    Its front-end functions are :
        elasticA3 ...... computes the coefficient A3 from the SOECs and TOECs
@@ -15,292 +15,25 @@
 #################################
 import ast
 import numpy as np
-from scipy.integrate import trapezoid
 from scipy.optimize import curve_fit, minimize_scalar, root
 import pandas as pd
-from pydislocdyn.utilities import Ncores, jit, usefortran, delta, hbar, kB, str2bool, OPTIONS, \
+from pydislocdyn.utilities import Ncores, usefortran, hbar, kB, str2bool, OPTIONS, \
     plt, fntsettings, AutoMinorLocator ## matplotlib stuff
 from pydislocdyn.elasticconstants import UnVoigt
 from pydislocdyn.dislocations import Dislocation, fourieruij_sincos, fourieruij_nocut, fourieruij_iso
 if Ncores>1:
     from joblib import Parallel, delayed
 if usefortran:
-    import pydislocdyn.subroutines as fsub
-    fourieruij_sincos = fsub.fourieruij_sincos
-
-@jit(nopython=True)
-def phonon(T,omega,q):
-    '''Compute a phonon distribution function. Requires input: temperature T in Kelvin, omega*q denotes the phonon energy over hbar.
-       A scale omega must be given separately, so that q is roughly of order 1. Otherwise 'overflow encountered in exp' errors might occur.'''
-    scale = hbar*omega/(kB*T)
-    return 1/(np.exp(scale*q)-1)
-
-if usefortran:
+    from pydislocdyn.subroutines import elastica3, phononwind_xx, phononwind_xy
     def elasticA3(C2, C3):
         '''Returns the tensor of elastic constants as it enters the interaction of dislocations with phonons. Required inputs are the tensors of SOEC and TOEC.'''
-        A3 = fsub.elastica3(C2,C3)
+        A3 = elastica3(C2,C3)
         return A3
 else:
-    @jit(nopython=True)
-    def elasticA3(C2, C3):
-        '''Returns the tensor of elastic constants as it enters the interaction of dislocations with phonons. Required inputs are the tensors of SOEC and TOEC.'''
-        A3 = C3.copy()
-        for i in range(3):
-            for ii in range(3):
-                for j in range(3):
-                    for jj in range(3):
-                        for k in range(3):
-                            for kk in range(3):
-                                A3[i,ii,j,jj,k,kk] += C2[i,ii,jj,kk]*delta[j,k] + C2[j,jj,ii,kk]*delta[i,k] + C2[ii,jj,k,kk]*delta[i,j]
-        return A3
-
-### this fct. needs dij for a fixed angle theta (no array in theta space)!
-@jit(nopython=True)
-def dragcoeff_iso_Bintegrand(prefactor,dij,poly):
-    '''Subroutine of dragcoeff_iso().'''
-    ## prefactor has shape (len(t),len(phi))
-    result1 = np.zeros(prefactor.shape)
-    prefacOnes = np.ones((len(prefactor)))
-    for k in range(3):
-        for kk in range(3):
-            for n in range(3):
-                for nn in range(3):
[…]
