+++ pydislocdyn/dislocations/steadystate.py
@@ -8,9 +8,33 @@
 #################################
 import numpy as np
 import sympy as sp
-from ..utilities import usefortran, jit, trapz, cumtrapz, delta, rotaround, artan, elbrak
+import pydislocdyn.dislocations.numba_subroutines as nsub
+from ..utilities import usefortran, delta, rotaround, artan
 if usefortran:
     from ..utilities import fsub
+    fourieruij_sincos = fsub.fourieruij_sincos
+    
+    def fourieruij_nocut(uij,phiX,sincos,Ntheta,phres):
+        '''Fourier transform of angular part of uij (needs result of subroutine fourieruij_sincos for sincos)'''
+        return fsub.fourieruij_nocut(np.moveaxis(uij,-1,0),phiX,sincos,Ntheta,phres)
+    
+    def computeEtot(uij, betaj, C2, Cv, phi):
+        '''Computes the self energy of a straight dislocation uij moving at velocity betaj.
+        Additional required input parameters are the 2nd order elastic constant tensor C2 (for the strain energy) and its velocity dependent shift Cv (for the kinetic energy),
+        as well as the integration angle phi inside the plane normal to the dislocation line.'''
+        return fsub.computeetot(np.moveaxis(uij,-1,0),betaj,C2,Cv,phi)
+    
+    def elbrak(A,B,elC):
+        '''Compute the bracket (A,B) := A.elC.B, where elC is a tensor of 2nd order elastic constants (potentially shifted by a velocity term or similar) and A,B are vectors.
+           All arguments are arrays, i.e. A and B have shape (3,Ntheta) where Ntheta is e.g. the number of character angles.'''
+        return np.moveaxis(fsub.elbrak(np.moveaxis(A,-1,0),np.moveaxis(B,-1,0),elC),0,-1)
+    
+    def elbrak1d(A,B,elC):
+        '''Compute the bracket (A,B) := A.elC.B, where elC is a tensor of 2nd order elastic constants (potentially shifted by a velocity term or similar) and A,B are vectors.
+           This function is similar to elbrak(), but its arguments do not depend on the character angle, i.e. A, B have shape (3).'''
+        return fsub.elbrak1d(A,B,elC)
+else:
+    from pydislocdyn.dislocations.numba_subroutines import fourieruij_sincos, fourieruij_nocut, computeEtot, elbrak, elbrak1d
 
 class StrohGeometry:
     '''This class computes several arrays to be used in the computation of a dislocation displacement gradient field for crystals using the integral version of the Stroh method.
@@ -116,7 +140,7 @@ class StrohGeometry:
             if r is not None:
                 self.uij = np.moveaxis(np.reshape(np.outer(1/r,self.uij),(len(r),3,3,self.Ntheta,len(self.phi))),0,-2)
         else:
-            self.uij = computeuij(beta, C2, self.Cv, self.b, self.M, self.N, self.phi, r=r, debug=debug)
+            self.uij = nsub.computeuij(beta, C2, self.Cv, self.b, self.M, self.N, self.phi, r=r, debug=debug)
         if abs(beta)<1e-15:
             self.uij_static = self.uij.copy()
         if debug:
@@ -144,7 +168,7 @@ class StrohGeometry:
         if usefortran:
             self.uk = np.moveaxis(fsub.computeuk(beta, C2, self.Cv, self.b, np.moveaxis(self.M,-1,0), np.moveaxis(self.N,-1,0), self.phi, r),0,-1)
         else:
-            self.uk = computeuij(beta, C2, self.Cv, self.b, self.M, self.N, self.phi, r=r, nogradient=True)
[…]
