+++ pydislocdyn/dislocations/__init__.py
@@ -2,7 +2,7 @@
 # PyDislocDyn
 # Author: Daniel N. Blaschke
 # Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-# Date: Apr. 3, 2024
+# Date: Apr. 3, 2024 - Sept. 13, 2025
 '''This submodule defines the StrohGeometry class to calculate the displacement field of a steady state dislocation
    aand other properties, as well as the Dislocation class which inherits from StrohGeometry
    and metal_props of polycrystal_averaging.py. As such, it is the most complete class to compute properties
@@ -10,5 +10,5 @@
    additional properties like limiting velocities of dislocations. We also define a function, readinputfile,
    which reads a PyDislocDyn input file and returns an instance of the Dislocation class.'''
 
-from .steadystate import StrohGeometry, computeuij_iso, fourieruij_sincos, fourieruij_nocut, fourieruij_iso
+from .steadystate import StrohGeometry, computeuij_iso, fourieruij_sincos, fourieruij_nocut, fourieruij_iso, elbrak, elbrak1d
 from .general import Dislocation, plotuij, readinputfile
