+++ manual/manual.tex
@@ -50,7 +50,7 @@ colorlinks=true,linkcolor=DarkBlue,citecolor=DarkGreen,urlcolor=Maroon]{hyperref
 		\end{flushright}\vspace{2em}}{}%
 	PyDislocDyn --- Code Manual}
 \author{Daniel N. Blaschke}
-\date{June 26, 2025}
+\date{September 17, 2025}
 
 \begin{document}
 \maketitle
@@ -78,7 +78,7 @@ Its main features summarize as:
 \item compute the steady-state displacement (gradient) field of a dislocation for arbitrary character angle in an arbitrary crystal geometry \cite{Bacon:1980,Hirth:1982}; a generalization to accelerating dislocations is in progress and so far has been implemented for pure screw dislocations \cite{Blaschke:2020acc} and pure edge dislocations \cite{Blaschke:2023acc}
 \item compute theoretical limiting velocities of dislocations in arbitrary slip systems and with arbitrary character angle \cite{Blaschke:2021vcrit,Blaschke:2020MD,Teutonico:1961,Barnett:1973b}
 \item compute the elastic strain energy and line tension for any dislocation \cite{Blaschke:2017lten}
-\item compute the dislocation drag coefficient $B$ from phonon wind for any dislocation at any velocity \cite{Blaschke:BpaperRpt,Blaschke:2018anis,Blaschke:2019fits,Blaschke:2019Bpap} or stress \cite{Blaschke:2019a,Blaschke:2021impact,Blaschke:2021temperature}.
+\item compute the dislocation drag coefficient $B$ from phonon wind for any dislocation at any velocity \cite{Alshits:1979,Alshits:1992,Blaschke:BpaperRpt,Blaschke:2018anis,Blaschke:2019fits,Blaschke:2019Bpap} or stress \cite{Blaschke:2019a,Blaschke:2021impact,Blaschke:2021temperature}.
 \end{itemize}
 
 \noindent
@@ -96,19 +96,20 @@ where it is distributed under the BSD-3 license, (Copyright, Triad National Secu
 The last version to support Python 2.7 was PyDislocDyn 1.2.0, Python 3.5 was supported up to version 1.2.3, Python 3.6 up to version 1.2.4, and Python 3.8 up to version 1.2.9.}.
 As of PyDislocDyn 1.3.1, required modules include Numpy $\ge1.19$, Scipy $\ge1.9$, Sympy $\ge1.9$, Matplotlib $\ge3.3$, Pandas $\ge1.3$ (including its optional dependency Jinja2), and mpmath (recommended optional dependency of mpmath: gmpy).
 \item Optional, but highly recommended in order to conduct drag coefficient calculations with reasonable speed are:
-Numba $\ge0.53$ (a just-in-time compiler for Python), Joblib $\ge1.1$ (for parallelization), and a Fortran compiler supported by numpy.f2py to compile the optional Fortran subroutines.
+a Fortran compiler supported by numpy.f2py to compile the optional Fortran subroutines, Joblib $\ge1.1$ (for parallelization), and 
+Numba $\ge0.53$ (a just-in-time compiler for Python; the Fortran subroutines are faster and thus preferred however).
 %To compile the optional Fortran subroutines, run `\verb|python -m numpy.f2py -c subroutines.f90 -m subroutines|'.\\
 %Optionally, the subroutines can also be compiled with OpenMP parallelization support by passing the appropriate options for your compiler. For example, with gfortran on Linux or MacOS with Python $\leq3.11$ compile with:
 %\\`\verb|python -m numpy.f2py --f90flags=-fopenmp -lgomp -c subroutines.f90|\\\verb|-m subroutines|';\\
[…]
