+++ manual/manual.bib
@@ -3,9 +3,29 @@
 	title = {{PyDislocDyn}},
 	url = {https://github.com/dblaschke-LANL/PyDislocDyn},
 	doi = {10.11578/dc.20180619.15},
-	version = {1.3.2},
+	version = {1.3.3},
 	year = {2018--2025},
 	reportNumber = {C18073}
+}
+@article{Alshits:1979,
+Author = {Al'shits, Vladimir I. and Mitlianskij, M. D. and Kotowski, R. K.},
+Title = {The phonon wind as a non-linear mechanism of dislocation dragging},
+Journal = {Arch. Mech.},
+Year = {1979},
+Volume = {31},
+Pages = {91-105},
+}
+@incollection{Alshits:1992,
+author = "Alshits, Vladimir I.",
+title = "The Phonon-Dislocation Interaction and its Role in Dislocation Dragging and Thermal Resistivity",
+editor = "Vladimir L. Indenbom and Jens Lothe",
+booktitle = "Elastic Strain Fields and Dislocation Mobility",
+publisher = "Elsevier",
+year = "1992",
+volume = "31",
+pages = "625-697",
+series = "Modern Problems in Condensed Matter Sciences",
+doi = "10.1016/B978-0-444-88773-3.50018-2",
 }
 @article{Bacon:1980,
 	author = "D. J. Bacon and D. M. Barnett and R. O. Scattergood",
