+++ readme.md
@@ -47,7 +47,7 @@ A helper function, pydislocdyn.utilities.compilefortranmodule(), is included to 
 * [joblib](https://joblib.readthedocs.io) >=1.1 (for parallelization),</br>
 * [numba](https://numba.pydata.org/) >=0.53 (for speedup via just-in-time compilation of some subroutines, although the Fortran subroutines are faster and thus preferred),</br>
 * [jupyter](https://jupyter.org/) to view and run the examples notebook
-* a recent version of LaTeX to build the manual (LA-UR-22-28074), pdf available at [doi:10.2172/1880452](https://doi.org/10.2172/1880452)
+* a recent version of LaTeX to build the manual (LA-UR-22-28074), PDF of Jan. 2025 version available at [doi:10.2172/1880452](https://doi.org/10.2172/1880452)
 
 ## Installation
 
