+++ manual/manual.bib
@@ -14,6 +14,7 @@ Journal = {Arch. Mech.},
 Year = {1979},
 Volume = {31},
 Pages = {91-105},
+url = {https://rcin.org.pl/dlibra/publication/90447},
 }
 @incollection{Alshits:1992,
 author = "Alshits, Vladimir I.",
