### 2025-09-22 ed4c52d
small code cleamup
