+++ pydislocdyn/subroutines.f90
@@ -1,4 +1,4 @@
-! Some subroutines meant to be compiled for phononwind.py and dislocations.py via f2py
+! Some subroutines meant to be compiled for pydislocdyn via f2py
 ! run 'python -m numpy.f2py -c subroutines.f90 -m subroutines' to use
 ! Author: Daniel N. Blaschke
 ! Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
