### 2025-10-02 968c9b7
prepare for near-future numpy deprecations, bump version
1.3.3 release candidate
