+++ pyproject.toml
@@ -9,7 +9,7 @@ packages = ['pydislocdyn','pydislocdyn.dislocations']
 
 [project]
 name = "PyDislocDyn"
-version = "1.3.3-beta"
+version = "1.3.3"
 authors = [
   { name="Daniel N. Blaschke", email="40176390+dblaschke-LANL@users.noreply.github.com" },
 ]
