### 2025-10-07 00f5d20
use threadpoolctl if available
reason: setting the environment variable for openmp threads
does not always work (e.g. when another module was previously
loaded that uses openmp); threadpoolctl fixes this issue
