### 2025-10-07 84f3e7e
compilefortranmodule(): new 'clean' option
to remove previously compiled module file(s)
and we now track compilation error codes in textfiles
fmoderror_py*.txt
