### 2025-10-07 f9f576d
ask to autocompile fortran module
upon importing pydislocdyn if it is missing or outdated.
Will not ask again if declined by user or compilation failed.
(compile manually by calling pydislocdyn.utilities.compilefortranmodule()
in this case)
