### 2025-10-07 f36c472
make auto-setting the number of openmp threads more robust
by setting the environment variable before importing numpy
(additionally we use threadpoolctl if it is installed)
