+++ pyproject.toml
@@ -9,7 +9,7 @@ packages = ['pydislocdyn','pydislocdyn.dislocations']
 
 [project]
 name = "PyDislocDyn"
-version = "1.3.3"
+version = "1.3.3+dev"
 authors = [
   { name="Daniel N. Blaschke", email="40176390+dblaschke-LANL@users.noreply.github.com" },
 ]
@@ -34,6 +34,7 @@ dependencies = [
 fast = [
   "numba>=0.53",
   "joblib>=1.1",
+  "threadpoolctl"
 ]
 
 [project.urls]
