### 2025-10-07 53ecb05
test suite: change order of imports
i.e. import pydislocdyn before numpy to set ompthreads first
if the user does not have threadpoolctl installed, this has
a good chance of working anyway
