### 2025-10-08 8dc5c2b
ompthreads is now a function, not a constant
i.e. it will always show the current number of OpenMP threads
(even if the user changed it with e.g. threadpoolctl)
if the Fortran subroutines are compiled with OpenMP support
(and 0 otherwise)
