### 2025-10-23 453b82d
fortran subroutines: address double-to-single conversion warnings
