### 2025-10-27 508ccab
fortran module: move all selected_real_kind calls into one place
