+++ pydislocdyn/subroutines.f90
@@ -2,16 +2,17 @@
 ! run 'python -m numpy.f2py -c subroutines.f90 -m subroutines' to use
 ! Author: Daniel N. Blaschke
 ! Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-! Date: July 23, 2018 - Oct. 23, 2025
+! Date: July 23, 2018 - Oct. 24, 2025
 
 subroutine version(versionnumber)
   integer, intent(out) :: versionnumber
-  versionnumber=20251023
+  versionnumber=20251024
 end subroutine version
 
 module parameters
 implicit none
 integer,parameter :: sel = selected_real_kind(10)
+integer,parameter :: selsm = selected_real_kind(6)  ! some memory-heavy subroutines use lower precision in favor of speed
 real(kind=sel), parameter :: hbar = 1.0545718d-34       ! reduced Planck constant
 real(kind=sel), parameter :: kB = 1.38064852d-23        ! Boltzmann constant
 real(kind=sel), parameter :: pi = (4.d0*atan(1.d0)) ! pi
@@ -32,16 +33,16 @@ subroutine parathesum(output,tcosphi,sqrtsinphi,tsinphi,sqrtcosphi,sqrtt,qv,delt
                       lenp,lent,lenph1,lentph)
 ! this wrapper parallelizes thesum() which is a subroutine of dragcoeff_iso_computepoly() in phononwind.py
 !$   Use omp_lib
+use parameters, only : selsm
 implicit none
-integer,parameter :: sel = selected_real_kind(6)
 
 integer, intent(in) :: lentph, lenph1, lenp, lent
-real(kind=sel), intent(in), dimension(lenph1) :: phi1, dphi1
-real(kind=sel), intent(in), dimension(lentph) :: mag, tcosphi, sqrtsinphi, tsinphi, sqrtcosphi, sqrtt
-real(kind=sel), intent(in), dimension(lentph,3) :: qv
-real(kind=sel), intent(in), dimension(3,3) :: delta1, delta2
-real(kind=sel), intent(in), dimension(3,3,3,3,3,3) :: A3
-real(kind=sel), intent(out), dimension(lentph,3,3,3,3) :: output
+real(kind=selsm), intent(in), dimension(lenph1) :: phi1, dphi1
+real(kind=selsm), intent(in), dimension(lentph) :: mag, tcosphi, sqrtsinphi, tsinphi, sqrtcosphi, sqrtt
+real(kind=selsm), intent(in), dimension(lentph,3) :: qv
+real(kind=selsm), intent(in), dimension(3,3) :: delta1, delta2
+real(kind=selsm), intent(in), dimension(3,3,3,3,3,3) :: A3
+real(kind=selsm), intent(out), dimension(lentph,3,3,3,3) :: output
 !$ integer :: i,j,k, nthreads
 output(:,:,:,:,:) = 0.0
 !$ call ompinfo(nthreads)
@@ -61,18 +62,18 @@ end subroutine parathesum
 
 subroutine thesum(output,tcosphi,sqrtsinphi,tsinphi,sqrtcosphi,sqrtt,qv,delta1,delta2,mag,A3,phi1,dphi1,lenph1,lentph)
 ! this is a subroutine of dragcoeff_iso_computepoly() in phononwind.py
+use parameters, only : selsm
 implicit none
-integer,parameter :: sel = selected_real_kind(6)
 integer :: i, j, k, l, ii, jj, kk, n, nn, m, p
 integer, intent(in) :: lentph, lenph1
-real(kind=sel), intent(in), dimension(lenph1) :: phi1, dphi1
[…]
