### 2025-10-27 63b978d
test suite: additional doc strings added
