### 2025-10-27 #9309 @editorialbot
[REVIEW]: PyDislocDyn: A Python code for calculating dislocation drag and other crystal properties
**Submitting author:** @dblaschke-LANL (<a href="http://orcid.org/0000-0001-5138-1462">Daniel N. Blaschke</a>)
**Repository:** https://github.com/dblaschke-LANL/PyDislocDyn
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 1.3.2
