### 2025-10-29 472fbcb
make test suite platform independent
