### 2025-10-29 78f8bdd
test suite: fix for win
