+++ manual/manual.bib
@@ -113,6 +113,7 @@ doi = {10.1007/978-1-4684-6066-7},
 	volume = "15a",
 	pages = "1056-1067",
 	year = "1960",
+	doi = {10.1515/zna-1960-1205},
 	url = "http://zfn.mpdl.mpg.de/data/Reihe_A/15/ZNA-1960-15a-1056.pdf",
 }
 @article{Graham:1968,
