### 2025-10-29 276bbc7
joss paper: fix an url/doi
