+++ joss_pydislocdyn.bib
@@ -92,7 +92,7 @@ doi = {10.1007/978-1-4684-6066-7},
 	volume = "15a",
 	pages = "1056-1067",
 	year = "1960",
-	url = "http://zfn.mpdl.mpg.de/data/Reihe_A/15/ZNA-1960-15a-1056.pdf",
+	doi = {10.1515/zna-1960-1205},
 }
 @article{Graham:1968,
 	author = "Graham, L. J. and Nadler, H. and Chang, Roger",
