+++ changelog.md
@@ -11,6 +11,8 @@ Features and improvements:
 Fix:
 
  - make auto-setting the number of openmp threads more robust: will use threadpoolctl if it is installed (new optional dependency)
+ - fix compiler warnings in the fortran subroutines
+ - make the test suite platform independent
 
 Other:
 
