### 2025-11-04 f96fe6c
nicer formatting in readme
