### 2025-11-04 0df99bc
updated joss paper
added a sentence at the beginning of the 'statement of need'
