### 2025-11-04 #9309 @dblaschke-LANL
Hi @gnzng , thanks for your feedback:
.) Statement of need:  I added an according new sentence at the beginning of that section.
.) Documentation: A link to the pdf of the Jan. 2025 version of the manual is already included in the readme. As for including it on github: I’ll make sure to include a pdf of the manual in the next release (which will be 1.3.4 and I plan to release it as soon as I've addressed all reviewer feedback on this joss submission).
.) Community docs: I will look into it and then post here again
.) Tests and CI: I don’t have experience with github actions and will have to first learn how to use them, i.e. I will look into it and then post here again.
.) README: done
.) Packaging: this is already stated in the readme file: ‘’’or (in order to install also the optional dependencies numba and joblib):
pip install ".[fast]" ‘’’
.) Add DOIs: the second doi has been included in the manual (see commit e2aa748 of Oct 29); the other two are wrong (pointing to a discussion of the book instead of the book and a completely different paper with the same title but different authors) and the robot suggesting them is mistaken, so nothing to do there
