+++ joss_pydislocdyn.bib
@@ -3,7 +3,7 @@
 	title = {{PyDislocDyn}},
 	url = {https://github.com/dblaschke-LANL/PyDislocDyn},
 	doi = {10.11578/dc.20180619.15},
-	version = {1.3.2},
+	version = {1.3.3},
 	year = {2018--2025},
 	reportNumber = {C18073}
 }
