### 2025-11-06 686695b
add github action for automated testing
