+++ .github/workflows/CI.yml
@@ -1,0 +1,50 @@
+name: CI
+
+on: push
+
+jobs:
+  RunTests:
+    runs-on: ${{ matrix.os }}
+
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+        python-version: ['3.13']
+        toolchain:
+          - {compiler: gcc, version: 'latest'}
+    steps:
+      - uses: actions/checkout@v3
+      - name: Setup Fortran Compiler
+        uses: fortran-lang/setup-fortran@v1
+        if: runner.os == 'MacOS'
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install .
+          pip install numpy scipy sympy matplotlib pandas jinja2 joblib threadpoolctl gmpy2 numba meson ninja
+      - name: Run test suite
+        run: yes 2>/dev/null | testing/test_pydislocdyn.py . --Ntheta=3 --Nbeta=5 --maxrec=1 --Nchunks=1 --verbose=true --runtests=all | tee test_output.log
+        if: runner.os != 'Windows'
+      - name: Run test suite windows
+        run: echo 'N' | python testing/test_pydislocdyn.py --Ntheta=3 --Nbeta=5 --maxrec=0 --Nchunks=1 --bccslip=112 --hcpslip=prismatic --verbose=true --runtests=all .
+        if: runner.os == 'Windows'
+      - name: Upload main log file as artifact
+        uses: actions/upload-artifact@v4
+        with:
+          name: test-log_${{runner.os}}
+          path: test_output.log
+        if: runner.os != 'Windows'
+      - name: Upload other log files as artifacts
+        uses: actions/upload-artifact@v4
+        with:
+          name: test-logfiles_${{runner.os}}
+          path: |
+            poly.log
+            dragiso.log
+            drag/dragsemi.log
+            LT_exp/LT.log
+
