### 2025-11-06 cbacd02
add info on feedback, how to contribute, etc.
