### 2025-11-06 #9309 @dblaschke-LANL
Hi @gnzng , I just addressed your remaining requests on the joss_paper branch, in particular:
.) Community docs: I added a short section at the end of the readme.md as well as a new CONTRIBUTING.md file linked from that section in the readme.md. (As I am currently the only maintainer/contributor, a code of conduct seems premature at this point)
.) automated tests have been added via github actions.

I think that was everything.
