### 2025-11-06 a00c3de
add link to wiki where we now host a pdf of the manual
