+++ CONTRIBUTING.md
@@ -1,0 +1,28 @@
+## Software Grant and Individual Contributor License Agreement (“Agreement”)
+
+Thank you for your interest in PyDislocDyn. In order to clarify the intellectual property license granted with Contributions from any person or entity, Triad National Security, LLC, as operator of Los Alamos National Laboratory (“LANL”) must have a Contributor License Agreement (CLA) on file that has been signed by each Contributor, indicating agreement to the license terms below.  This license is for your protection as a Contributor as well as the protection of LANL and the PyDislocDyn users; it does not change your rights to use your own Contributions for any other purpose.
+
+If you have not already done so, please complete and sign, then scan and email a pdf file of this Agreement to software@lanl.gov.  Alternatively, you may send it by facsimile to LANL at 505-665-3125.  If necessary, send an original, signed Agreement to LANL, P.O. Box 1663, MS C333, Attn. Contracts Management Team, Los Alamos, NM  87545. 
+
+Please read this document carefully before signing and keep a copy for your records.
+
+You accept and agree to the following terms and conditions for Your present and future Contributions submitted to LANL.  In return, LANL shall not use your Contributions in a way that is contrary to the public benefit or inconsistent with its goals and objectives as a federal laboratory.  Except for the license granted herein to Triad National Security, LLC and recipients of software distributed by LANL, You reserve all right, title, and interest in and to Your Contributions.
+
+1.	 Definitions.  “You” (or “Your”) shall mean the copyright owner or legal entity authorized by the copyright owner that is making this Agreement with Triad National Security, LLC.  For legal entities, the entity making a Contribution and all other entities that control, are controlled by, or are under common control with that entity are considered to be a single Contributor.  For the purposes of this definition, “control” means (i) the power, direct or indirect, to cause the direction or management of such entity, whether by contract or otherwise, or (ii) ownership of fifty percent (50%) or more of the outstanding shares, or (iii) beneficial ownership of such entity.
+
+“Contribution” shall mean any original work of authorship, including any modifications or additions to an existing work, that are intentionally submitted by You to Triad National Security, LLC for inclusion in, or documentation of (insert software name)  (the "Work"). For the purposes of this definition, "submitted" means any form of electronic, verbal, or written communication sent to Triad National Security, LLC or its representatives, including but not limited to communication on electronic mailing lists, source code control systems, and issue tracking systems that are managed by, or on behalf of, Triad National Security, LLC for the purpose of discussing and improving the Work, but excluding communication that is conspicuously marked or otherwise designated in writing by You as "Not a Contribution."
+
+2.	 Grant of Copyright License.  Subject to the terms and conditions of this Agreement, You hereby grant to Triad National Security, LLC and to recipients of (insert software name) a perpetual, worldwide, non-exclusive, no-charge, royalty-free, irrevocable copyright license to reproduce, prepare derivative works of, publicly display, publicly perform, sublicense, and distribute Your Contributions and such derivative works.
+
+3.	You represent that You are legally entitled to grant the above-referenced Patent License and the Copyright License.  If your employer(s) has rights to intellectual property that you create that includes your Contributions, you represent that you have received permission to make Contributions on behalf of that employer, that your employer has waived such rights for your Contributions to Triad National Security, LLC, or that your employer has executed a separate Corporate Agreement with Triad National Security, LLC.
+
+4.	You represent that each of Your Contributions is Your original creation (see section 7 for submission on behalf of others).  You represent that Your Contribution submissions include complete details of any third-party license or other restriction (including, but not limited to, related patents and trademarks) of which you are personally aware and which are associated with any part of Your Contributions.
+
[…]
