### 2025-11-06 #9309 @dblaschke-LANL
Hi @gnzng regarding the documentation request:
A pdf of the manual has now been included on the newly created wiki page (linked from the readme).
