### 2025-11-07 #9309 @gnzng
I just see you merged `joss_paper` into `master`. All the additions to the code, which should be reviewed has to be on the default branch, in this case `master`. The paper related info `paper.md` and `.bib` can stay in the `joss_paper` branch.

Also somehow the testing CI implementation is not executed.
