### 2025-11-07 #9309 @dblaschke-LANL
Hi @gnzng ,

all changes are in master, i.e. the only difference between the master branch and the joss_paper branch at this point is the paper.md, bib-file and github-action to compile paper.md; everything else is in sync:
$ git diff --numstat joss_paper
0       24      .github/workflows/draft-pdf.yml
0       975     joss_pydislocdyn.bib
0       109     paper.md

The github action to run the test suite (CI.yml) has been configured to run whenever *.py, *.f90 or the toml file change (i.e. it will not run for documentation updates); last run (with all updates in the code) was
https://github.com/dblaschke-LANL/PyDislocDyn/actions/runs/19151304679
and it was successful.

Hope this clarifies things!
