### 2025-11-13 4bc4512
rename license file, bump version etc.
