+++ CONTRIBUTING.md
@@ -10,7 +10,7 @@ You accept and agree to the following terms and conditions for Your present and 
 
 1.	 Definitions.  “You” (or “Your”) shall mean the copyright owner or legal entity authorized by the copyright owner that is making this Agreement with Triad National Security, LLC.  For legal entities, the entity making a Contribution and all other entities that control, are controlled by, or are under common control with that entity are considered to be a single Contributor.  For the purposes of this definition, “control” means (i) the power, direct or indirect, to cause the direction or management of such entity, whether by contract or otherwise, or (ii) ownership of fifty percent (50%) or more of the outstanding shares, or (iii) beneficial ownership of such entity.
 
-“Contribution” shall mean any original work of authorship, including any modifications or additions to an existing work, that are intentionally submitted by You to Triad National Security, LLC for inclusion in, or documentation of (insert software name)  (the "Work"). For the purposes of this definition, "submitted" means any form of electronic, verbal, or written communication sent to Triad National Security, LLC or its representatives, including but not limited to communication on electronic mailing lists, source code control systems, and issue tracking systems that are managed by, or on behalf of, Triad National Security, LLC for the purpose of discussing and improving the Work, but excluding communication that is conspicuously marked or otherwise designated in writing by You as "Not a Contribution."
+  	 “Contribution” shall mean any original work of authorship, including any modifications or additions to an existing work, that are intentionally submitted by You to Triad National Security, LLC for inclusion in, or documentation of (insert software name)  (the "Work"). For the purposes of this definition, "submitted" means any form of electronic, verbal, or written communication sent to Triad National Security, LLC or its representatives, including but not limited to communication on electronic mailing lists, source code control systems, and issue tracking systems that are managed by, or on behalf of, Triad National Security, LLC for the purpose of discussing and improving the Work, but excluding communication that is conspicuously marked or otherwise designated in writing by You as "Not a Contribution."
 
 2.	 Grant of Copyright License.  Subject to the terms and conditions of this Agreement, You hereby grant to Triad National Security, LLC and to recipients of (insert software name) a perpetual, worldwide, non-exclusive, no-charge, royalty-free, irrevocable copyright license to reproduce, prepare derivative works of, publicly display, publicly perform, sublicense, and distribute Your Contributions and such derivative works.
 
