### 2025-11-14 72212fb
fix formatting in license file
