### 2025-11-14 17f151e
rename license file, bump version
