### 2025-11-15 eb082f6
test suite: fix passing an option
