### 2025-11-22 f903470
use pathlib for paths
