+++ testing/test_pydislocdyn.py
@@ -9,13 +9,13 @@
 import os
 import sys
 import subprocess
-import glob
+import pathlib
 import difflib
 import lzma
-dir_path = os.path.realpath(os.path.join(os.path.dirname(os.path.realpath(__file__)),os.pardir))
+dir_path = str(pathlib.Path(__file__).resolve().parents[1])
 if dir_path not in sys.path:
     sys.path.append(dir_path)
-dir_path = os.path.join(dir_path,'pydislocdyn')
+dir_path = pathlib.Path(__file__).resolve().parents[1] / 'pydislocdyn'
 
 from pydislocdyn.metal_data import fcc_metals, bcc_metals, hcp_metals, tetr_metals, \
     ISO_l, c111, all_metals
@@ -77,7 +77,8 @@ def printtestresult(success,countfails=0):
 
 def readfile(fname):
     '''reads a text file (or xz compressed text file) and returns a list of its lines'''
-    if fname[-4:] == '.tex' and not os.path.isfile(fname):
+    fname = str(fname)
+    if fname[-4:] == '.tex' and not pathlib.Path(fname).is_file():
         fname = fname[:-4]+'.txt' ## allow comparing to old versions of pydislocdyn
     if fname[-3:] == '.xz':
         with lzma.open(fname,"rt") as f1:
@@ -123,9 +124,9 @@ def runscript(scriptname,args,logfname):
     '''Run script "scriptname" as a subprocess passing a list of command line arguments "args" and saving its stdout to a file "logfname"'''
     out = -1
     with open(logfname, 'w', encoding="utf8") as logfile:
-        command = [os.path.join(dir_path,scriptname)]
+        command = [pathlib.Path(dir_path,scriptname)]
         if sys.platform=='win32':
-            command = ["python",os.path.join(dir_path,scriptname)]
+            command = ["python",pathlib.Path(dir_path,scriptname)]
         with subprocess.Popen(command+args, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True) as subproc:
             for line in subproc.stdout:
                 sys.stdout.write(line)
@@ -136,7 +137,8 @@ def runscript(scriptname,args,logfname):
 
 if __name__ == '__main__':
     tests_avail=['all', 'aver', 'dragiso', 'drag', 'LT', 'acc', 'misc', 'obj']
-    cwd = os.getcwd()
+    cwd =pathlib.Path.cwd()
+    tmppydislocdyn = pathlib.Path("temp_pydislocdyn")
     no_failed_tests = 0
     if len(sys.argv) > 1:
         oldglobals = globals().copy()
@@ -145,7 +147,7 @@ if __name__ == '__main__':
         old, kwargs = parse_options(sys.argv[1:],OPTIONS,globals(),starthelpwith=starthelpwith)
         if len(old)==0:
             raise ValueError("missing one argument: folder containing old results")
-        old = old[0]
+        old = pathlib.Path(old[0])
         phononwind_opts.update(kwargs)
         NEWopts = globals().keys()-oldglobals.keys() ## pass options which we haven't previously defined but that the user has set
         LTopts = {i:globals()[i] for i in NEWopts if i in OPTIONS_LT.keys()}
@@ -154,7 +156,7 @@ if __name__ == '__main__':
         dragopts = [f" --{i}={j}" for i,j in dragopts.items()]
     else:
         raise ValueError("missing one argument: folder containing old results")
-    if os.path.exists(old):
+    if old.exists():
         print(f"comparing to {old}\n")
     else:
         raise ValueError(f"folder {old} does not exist")
@@ -204,7 +206,7 @@ if __name__ == '__main__':
[…]
