### 2025-11-22 e91e111
move all unit tests to pytest
and generalize them a little
