+++ testing/test_pydislocdyn.py
@@ -2,7 +2,7 @@
 # test suite for PyDislocDyn
 # Author: Daniel N. Blaschke
 # Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-# Date: Mar. 6, 2023 - Nov. 15, 2025
+# Date: Mar. 6, 2023 - Nov. 22, 2025
 '''This script implements regression testing for PyDislocDyn. Required argument: 'folder' containing old results.
    (To freshly create a folder to compare to later, run from within an empty folder with argument 'folder' set to '.')
    For additional options, call this script with '--help'.'''
@@ -20,8 +20,7 @@ dir_path = pathlib.Path(__file__).resolve().parents[1] / 'pydislocdyn'
 from pydislocdyn.metal_data import fcc_metals, bcc_metals, hcp_metals, tetr_metals, \
     ISO_l, c111, all_metals
 from pydislocdyn.utilities import parse_options, str2bool, isclose, compare_df
-from pydislocdyn import read_2dresults, Ncores, Voigt, UnVoigt, strain_poly, writeallinputfiles, \
-    readinputfile, convert_SOECiso, convert_TOECiso, Dislocation, roundcoeff
+from pydislocdyn import read_2dresults, Ncores, Voigt, strain_poly, writeallinputfiles, readinputfile
 from pydislocdyn.linetension_calcs import OPTIONS as OPTIONS_LT
 from pydislocdyn.dragcoeff_semi_iso import OPTIONS as OPTIONS_drag
 import numpy as np ## import pydislocdyn first as it will set the openmp thread number
@@ -136,7 +135,7 @@ def runscript(scriptname,args,logfname):
     return out
 
 if __name__ == '__main__':
-    tests_avail=['all', 'aver', 'dragiso', 'drag', 'LT', 'acc', 'misc', 'obj']
+    tests_avail=['all', 'aver', 'dragiso', 'drag', 'LT', 'acc', 'misc']
     cwd =pathlib.Path.cwd()
     tmppydislocdyn = pathlib.Path("temp_pydislocdyn")
     no_failed_tests = 0
@@ -449,44 +448,6 @@ if __name__ == '__main__':
                 Parallel(n_jobs=Ncores)(delayed(maincomputations)(sym) for sym in crystalsyms)
             else:
                 [maincomputations(sym) for sym in crystalsyms]
-            print("running some unit tests ...")
-            for X in metal_list:
-                Y[X].sumofsounds = 0
-                for v in ([1,0,0],[0,1,0],[0,0,1]):
-                    sound = Y[X].computesound(v)
-                    if len(sound)==2:
-                        Y[X].sumofsounds += 2*min(sound)**2+max(sound)**2
-                    else:
-                        Y[X].sumofsounds += sum(np.array(sound)**2)
-                Y[X].C2tracerho = np.trace(np.trace(UnVoigt(Y[X].C2),axis1=1,axis2=2))/Y[X].rho
-                if not np.isclose(Y[X].C2tracerho,Y[X].sumofsounds): ## see Fitzgerald 1967 for details on this relation
-                    print(f"invariant sum of sound speeds unit test failed for {X}: {Y[X].C2tracerho=}, {Y[X].sumofsounds=}")
-                    success = False
-                if Y[X].Zener is not None:
-                    Y[X].AL = Y[X].anisotropy_index()
-                    Y[X].AL_Z = np.sqrt(5)*np.log((2+3*Y[X].Zener)*(3+2*Y[X].Zener)/(25*Y[X].Zener))
[…]
