+++ testing/test_unit.py
@@ -1,0 +1,139 @@
+#!/usr/bin/env python3
+# test suite for PyDislocDyn
+# Author: Daniel N. Blaschke
+# Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
+# Date: Mar. 6, 2023 - Nov. 22, 2025
+'''This script implements several unit tests for PyDislocyn meant to be called by pytest.'''
+import os
+import sys
+import pathlib
+dir_path = str(pathlib.Path(__file__).resolve().parents[1])
+if dir_path not in sys.path:
+    sys.path.append(dir_path)
+testpath = pathlib.Path(__file__).resolve().parents[0]
+tmppydislocdyn = pathlib.Path(testpath,"temp_pydislocdyn")
+tmppydislocdyn.mkdir(exist_ok=True)
+
+import pydislocdyn as pydis
+import numpy as np
+import sympy as sp
+
+def initialize_metals(metal_list=None):
+    '''writes all input files to subfolder "temp_pydislocdyn", then reads a subset defined by the keywords in "metal_list",
+       and returns a dictionary of the resulting metal_props class instances.'''
+    os.chdir(tmppydislocdyn)
+    pydis.writeallinputfiles()
+    Y = {}
+    if metal_list is None:
+        metal_list = sorted(tmppydislocdyn.glob("*"))
+    for X in metal_list:
+        tmpY = pydis.crystals.readinputfile(X)
+        Y[tmpY.name] = tmpY
+    return Y
+
+def initialize_dislocs(metal_list=None,Ntheta=2):
+    '''writes all input files to subfolder "temp_pydislocdyn", then reads a subset defined by the keywords in "metal_list",
+       and returns a dictionary of the resulting Dislocation class instances.'''
+    os.chdir(tmppydislocdyn)
+    pydis.writeallinputfiles()
+    Y = {}
+    if metal_list is None:
+        metal_list = sorted(tmppydislocdyn.glob("*"))
+    for X in metal_list:
+        tmpY = pydis.readinputfile(X,Ntheta=Ntheta)
+        Y[tmpY.name] = tmpY
+    return Y
+
+def test_sound(metal_list=None):
+    '''verify properties of calculated sound speeds'''
+    Y = initialize_metals(metal_list=metal_list)
+    for X in Y:
+        Y[X].sumofsounds = 0
+        for v in ([1,0,0],[0,1,0],[0,0,1]):
+            sound = Y[X].computesound(v)
+            if len(sound)==2:
+                Y[X].sumofsounds += 2*min(sound)**2+max(sound)**2
+            else:
+                Y[X].sumofsounds += sum(np.array(sound)**2)
+        Y[X].C2tracerho = np.trace(np.trace(pydis.UnVoigt(Y[X].C2),axis1=1,axis2=2))/Y[X].rho
+        assert np.isclose(Y[X].C2tracerho,Y[X].sumofsounds) ## see Fitzgerald 1967 for details on this relation
+
+def test_elasticconstants(metal_list=None,Ntheta=2):
+    '''verify properties of elastic constants of anisotropic crystals'''
+    Y = initialize_dislocs(metal_list=metal_list,Ntheta=Ntheta)
+    for X in Y:
[…]
