### 2025-11-22 03e100b
Modernize1 (#1)
* use pathlib for paths
* move all unit tests to pytest and generalize them a little
