### 2025-12-02 31d77f3
joss paper: shortened first sentence
