### 2025-12-02 #9309 @dblaschke-LANL
@lucydot thanks for the suggestions! I have shortened the first sentence accordingly. Indeed, it does not only apply to metals, so that statement can go. Also, the phrase 'calculations for basic research' seemed unnecessary on second thought and the features section already states that calculations can be done for any crystal symmetry (2nd bullet), so I did not see the need to modify the features section further.
