### 2025-12-08 6c43907
fix wrong isbn in .bib (no was an issn)
