+++ joss_pydislocdyn.bib
@@ -817,7 +817,7 @@ author = {Tawqeer Nasir Tak and Aditya Prakash and Indradev Samajdar and Ahmed A
 @article{Bertin:2024,
 author = {Bertin, N. and Bulatov, Vasily V. and Zhou, Fei},
 doi = {10.1038/s41524-024-01378-4},
-isbn = {2057-3960},
+issn = {2057-3960},
 journal = {npj Computational Materials},
 number = {1},
 pages = {192},
