### 2025-12-12 d427c8c
update CI script for github actions
