+++ .github/workflows/CI.yml
@@ -14,7 +14,15 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ['3.13']
+        python-version: ["3.9", "3.14"]
+        exclude:
+          - os: macos-latest
+            python-version: '3.9'
+        include:
+          - python-version: '3.9'
+            module-versions: "numpy==1.21 scipy==1.9 sympy==1.9 matplotlib==3.3 pandas==1.3 jinja2 joblib==1.1 threadpoolctl gmpy2 numba==0.53 meson ninja pytest setuptools==69"
+          - python-version: '3.14'
+            module-versions: "numpy scipy sympy matplotlib pandas jinja2 joblib threadpoolctl gmpy2 numba meson ninja pytest"
         toolchain:
           - {compiler: gcc, version: 'latest'}
     steps:
@@ -29,8 +37,8 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
+          pip install ${{ matrix.module-versions }}
           pip install .
-          pip install numpy scipy sympy matplotlib pandas jinja2 joblib threadpoolctl gmpy2 numba meson ninja pytest
       - name: Run regression test suite
         run: yes 2>/dev/null | testing/test_pydislocdyn.py . --Ntheta=3 --Nbeta=5 --maxrec=1 --Nchunks=1 --verbose=true --runtests=all | tee test_output.log
         if: runner.os != 'Windows'
@@ -38,21 +46,21 @@ jobs:
         run: echo 'N' | python testing/test_pydislocdyn.py --Ntheta=3 --Nbeta=5 --maxrec=0 --Nchunks=1 --bccslip=112 --hcpslip=prismatic --verbose=true --runtests=all .
         if: runner.os == 'Windows'
       - name: Run unit tests
-        run: yes 2>/dev/null | pytest testing/
+        run: yes 2>/dev/null | pytest -s testing/
         if: runner.os != 'Windows'
       - name: Run unit tests windows
-        run: echo 'N' | pytest testing/
+        run: echo 'N' | pytest -s testing/
         if: runner.os == 'Windows'
       - name: Upload main log file as artifact
         uses: actions/upload-artifact@v4
         with:
-          name: test-log_${{runner.os}}
+          name: test-log_${{runner.os}}_${{matrix.python-version}}
           path: test_output.log
         if: runner.os != 'Windows'
       - name: Upload other log files as artifacts
         uses: actions/upload-artifact@v4
         with:
-          name: test-logfiles_${{runner.os}}
+          name: test-logfiles_${{runner.os}}_${{matrix.python-version}}
           path: |
             poly.log
             dragiso.log
