### 2025-12-20 47419a3
Update paper.md
