### 2026-01-05 25158df
Update readme.md
add ref. to joss paper
