### 2026-02-13 8db1f64
start porting options to argparse, LT complete
