### 2026-02-13 4747a4c
partial argparse port of phononwind calcs; frontend scripts incomplete
