+++ pydislocdyn/dragcoeff_iso.py
@@ -7,24 +7,6 @@
    it is not meant to be used as a module.
    The script takes as (optional) arguments either the names of PyDislocDyn input files or keywords for
    metals that are predefined in metal_data.py, falling back to all available if no argument is passed.
-
-Additional options (can be set on the commandline with syntax --keyword=value):
-   --skip_plots - default=False, set to True to skip generating plots from the results
-   --use_exp_Lame - default=True, if using data from metal_data, choose between experimentally determined Lame and Murnaghan constants (default) 
-                                  or analytical averages of SOEC and TOEC (use_exp_Lame = False)
-   --allplots - default=False, set to True to show more B_of_sigma plots for each metal
-
-Choose various resolutions and other parameters:
-    --Ntheta (default:2) - number of angles between Burgers vector and dislocation line (minimum 2, i.e. pure edge and pure screw)
-    --Nbeta (default:99) - number of velocities to consider ranging from minb to maxb (as fractions of transverse sound speed)
-    --minb (default:0.01)
-    --maxb (default:0.99)
-    
-    --modes - phonons to include ('TT'=pure transverse, 'LL'=pure longitudinal, 'TL'=L scattering into T, 'LT'=T scattering into L, 
-                                  'mix'=TL+LT, 'all'=sum of all four = default)
-    --Nphi - resolution of polar angle in Fourier space; keep this an even number for higher accuracy (because we integrate over 
-                pi-periodic expressions in some places and phi ranges from 0 to 2pi)
-    --phononwind_opts - pass additional options to the phononwind subroutine (formatted as a dictionary)
 '''
 #################################
 import sys
@@ -38,55 +20,39 @@ if dir_path not in sys.path:
     sys.path.append(dir_path)
 ##
 import pydislocdyn.metal_data as data
-from pydislocdyn.utilities import printthreadinfo, parse_options, showoptions, Ncores, read_2dresults, \
+from pydislocdyn.utilities import printthreadinfo, separate_options, showoptions, Ncores, read_2dresults, \
     plt, fntsettings, AutoMinorLocator, pd ## matplotlib stuff
 from pydislocdyn.dislocations import readinputfile
 from pydislocdyn.phononwind import phonondrag, B_of_sigma, OPTIONS, init_drag_parser
-
-Ntheta = 2
-Nbeta = 99
-minb = 0.01
-maxb = 0.99
-modes = 'all'
-skip_plots = False
-use_exp_Lame = True
-allplots = False
-NT = 1 # number of temperatures between baseT and maxT (WARNING: implementation of temperature dependence is incomplete!)
-constantrho = False ## set to True to override thermal expansion coefficient and use alpha_a = 0 for T > baseT
-increaseTby = 300 # so that maxT=baseT+increaseTby (default baseT=300 Kelvin, but may be overwritten by an input file below)
-beta_reference = 'base'  ## define beta=v/ct, choosing ct at baseT ('base') or current T ('current') as we increase temperature
-Nphi = 50
-phononwind_opts = {} ## 
 
 parser = init_drag_parser(usage=f"\n{sys.argv[0]} <options> <inputfile(s)>\n\n",description=f"{__doc__}\n")
 parser.add_argument('-Ntheta','--Ntheta', type=int, default=2, help="""number of angles between Burgers vector and dislocation line (minimum 2, i.e. pure edge and pure screw)""")
 
 #########
[…]
