### 2026-02-14 6bec4c1
finish porting the frontend scripts for using argparse
