+++ pydislocdyn/dragcoeff_semi_iso.py
@@ -7,36 +7,6 @@
 it is not meant to be used as a module.
 The script takes as (optional) arguments either the names of PyDislocDyn input files or keywords for
 metals that are predefined in metal_data.py, falling back to all available if no argument is passed.
-
-Additional options (can be set on the commandline with syntax --keyword=value):
-   --skip_plots - default=False, set to True to skip generating plots from the results
-   --use_exp_Lame - default=True, if set to True, experimental values (where available) are taken for the Lame constants, 
-                   isotropic phonon spectrum, and sound speeds missing values (such as Mo, Zr, or all if use_exp_Lame=False) 
-                   are supplemented by Hill averages, or for cubic crystals the 'improved average' (see 'polycrystal_averaging.py')
-   --use_iso - default=False, set to True to calculate using isotropic elastic constants from metal_data (no effect if input files are used)
-   --allplots - default=False, set to True to show more B_of_sigma plots for each metal
-
-Choose various resolutions and other parameters:
-    --Ntheta (default:21) - number of dislocation character angles between 0 and pi/2 (minimum 2, 
-                            i.e. pure edge and pure screw), if the range -pi/2--pi/2 is required the number of angles
-                            is increased to 2*Ntheta-1
-    --Nbeta (default:99) - number of velocities to consider ranging from minb to maxb (as fractions of transverse sound speed)
-    --minb (default:0.01)
-    --maxb (default:0.99)
-    
-    --modes - phonons to include ('TT'=pure transverse, 'LL'=pure longitudinal, 'TL'=L scattering into T, 'LT'=T scattering into L, 
-                                  'mix'=TL+LT, 'all'=sum of all four = default)
-    --Nphi - resolution of polar angle in Fourier space; keep this an even number for higher accuracy (because we integrate over 
-                pi-periodic expressions in some places and phi ranges from 0 to 2pi)
-    --phononwind_opts - pass additional options to the phononwind subroutine (formatted as a dictionary)
-
-Choose among predefined slip systems when using metal_data.py (see that file for details):
-   --bccslip - allowed values: '110', '112', '123', 'all' (for all three, default)
-   --hcpslip - allowed values: 'basal', 'prismatic', 'pyramidal', 'all' (for all three,default)
-
-   --skiptransonic - if True (default) will skip phononwind calcs. for velocities above the lowest limiting velocity on a per character angle 
-           basis, filling in the blanks with np.inf; Note: this will speed up calculations by avoiding slow converging drag calcs 
-           near divergences of the dislocation field
 '''
 #################################
 import sys
@@ -51,43 +21,15 @@ if dir_path not in sys.path:
     sys.path.append(dir_path)
 ##
 from pydislocdyn import metal_data as data
-from pydislocdyn.utilities import ompthreads, printthreadinfo, parse_options, showoptions, str2bool, Ncores, Ncpus, read_2dresults, \
+from pydislocdyn.utilities import ompthreads, printthreadinfo, separate_options, str2bool, Ncores, Ncpus, read_2dresults, \
     plt, fntsettings, AutoMinorLocator, gridspec, make_axes_locatable, pd ## matplotlib stuff
 from pydislocdyn.dislocations import readinputfile
[…]
