### 2026-02-14 41ef1d4
clean up some remnants and fix test suite for new version
