### 2026-02-14 6942ca4
remove old OPTIONS from front-end scripts
