### 2026-02-15 a4ca04e
add missing defaults
