### 2026-02-15 4207541
ported some regression tests to pytest and some small fixes
