+++ testing/test_regression.py
@@ -1,0 +1,271 @@
+#!/usr/bin/env python3
+# test suite for PyDislocDyn
+# Author: Daniel N. Blaschke
+# Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
+# Date: Mar. 6, 2023 - Feb. 15, 2026
+'''This script implements regression testing for PyDislocDyn. Required argument: 'folder' containing old results.
+   (To freshly create a folder to compare to later, run from within an empty folder with argument 'folder' set to '.')
+   For additional options, call this script with '--help'.'''
+import os
+import sys
+import subprocess
+import pathlib
+import difflib
+import lzma
+dir_path = str(pathlib.Path(__file__).resolve().parents[1])
+if dir_path not in sys.path:
+    sys.path.append(dir_path)
+dir_path = pathlib.Path(__file__).resolve().parents[1] / 'pydislocdyn'
+
+from pydislocdyn.metal_data import fcc_metals, bcc_metals, hcp_metals, tetr_metals, \
+    ISO_l, c111, all_metals
+from pydislocdyn.utilities import str2bool, isclose, compare_df
+from pydislocdyn import read_2dresults, Ncores, Voigt, strain_poly, writeallinputfiles, readinputfile
+import numpy as np ## import pydislocdyn first as it will set the openmp thread number
+import sympy as sp
+import pandas as pd
+if Ncores>1:
+    from joblib import Parallel, delayed
+    
+cwd =pathlib.Path.cwd()
+
+defaults_for_testing = {'P':0, ## pressure in strain_poly test
+                        'volpres':False, ## set to True to compute volume preserving version of the strains
+                        'fastapprox':True, ## set to False to include terms that are (close to) zero in acc_screw disloc. field
+                        'vRF_resolution':50,
+                        'vRF_fast':True}
+
+def readfile(fname):
+    '''reads a text file (or xz compressed text file) and returns a list of its lines'''
+    fname = str(fname)
+    if fname[-4:] == '.tex' and not pathlib.Path(fname).is_file():
+        fname = fname[:-4]+'.txt' ## allow comparing to old versions of pydislocdyn
+    if fname[-3:] == '.xz':
+        with lzma.open(fname,"rt") as f1:
+            f1lines = f1.readlines()
+    else:
+        with open(fname,"r", encoding="utf8") as f1:
+            f1lines = f1.readlines()
+    if fname[-4:] == '.tex' or fname[-4:] == '.txt': ## ignore expected changes in LaTeX code
+        f1lines = [x for x in f1lines if 'tabular' not in x and 'table' not in x and 'averages' not in x]
+    return f1lines
+
+def removewhitespace(somestring):
+    '''replaces excessive whitespace with ' ' in the middle of a string, and removes leading/trailing whitespace (space and tab characters only)'''
+    string1 = somestring.strip()
+    string2 = string1.replace("\t"," ").replace("  "," ")
+    while string2!=string1:
+        string1 = string2
+        string2 = string1.replace("\t"," ").replace("  "," ")
+    return string2
+
+def diff(f1,f2,verbose=True):
+    '''Compares two text files'''
+    f1lines = readfile(f1)
+    f1lines = [removewhitespace(i) for i in f1lines]
+    f2lines = readfile(f2)
+    f2lines = [removewhitespace(i) for i in f2lines]
+    thediff = difflib.unified_diff(f1lines, f2lines, fromfile=str(f1),tofile=str(f2),lineterm="",n=0)
+    equal = f1lines==f2lines
[…]
