### 2026-02-15 1db757c
port the remaining regression tests to pytest
