+++ testing/test_regression.py
@@ -29,11 +29,7 @@ if Ncores>1:
     
 cwd =pathlib.Path.cwd()
 
-defaults_for_testing = {'P':0, ## pressure in strain_poly test
-                        'volpres':False, ## set to True to compute volume preserving version of the strains
-                        'fastapprox':True, ## set to False to include terms that are (close to) zero in acc_screw disloc. field
-                        'vRF_resolution':50,
-                        'vRF_fast':True}
+########## define helper functions for the tests below
 
 def readfile(fname):
     '''reads a text file (or xz compressed text file) and returns a list of its lines'''
@@ -130,14 +126,16 @@ def expand_slipsystems(metals,bccslip='all',hcpslip='all'):
             out.append(X)
     return out
 
-def test_aver(old=None,skip_calcs=False,verbose=False):
+########## tests:
+
+def test_aver(old=None,new=cwd,skip_calcs=False,verbose=False):
     '''implements regression tests for frontend script polycrystal_averaging.py'''
-    testfolder = cwd/"regressiontests"
+    testfolder = new/"regressiontests"
     testfolder.mkdir(exist_ok=True)
     if old is None:
         old = testfolder
     else:
-        old = cwd / old
+        old = new / old
     if verbose:
         print(f"{testfolder=}, baseline folder={old}")
     os.chdir(testfolder)
@@ -148,17 +146,17 @@ def test_aver(old=None,skip_calcs=False,verbose=False):
     else: print("skipping test 'aver' as requested")
     print(f"checking {fname}")
     assert diff(pathlib.Path(old,fname),pathlib.Path(testfolder,fname),verbose=verbose)
-    os.chdir(cwd)
+    os.chdir(new)
 
-def test_dragiso(old=None,skip_calcs=False,verbose=False,metals='Cu Fe',**kwargs):
+def test_dragiso(old=None,new=cwd,skip_calcs=False,verbose=False,metals='Cu Fe',**kwargs):
     '''implements regression tests for isotropic phonon drag calculations via frontend script dragcoeff_iso.py,
        where folder "old" contains the baseline results; set to "None" to initialize a new baseline.'''
-    testfolder = cwd/"regressiontests"
+    testfolder = new/"regressiontests"
     testfolder.mkdir(exist_ok=True)
     if old is None:
         old = testfolder
     else:
-        old = cwd / old
+        old = new / old
     if verbose:
         print(f"{testfolder=}, baseline folder={old}")
     os.chdir(testfolder)
@@ -185,17 +183,17 @@ def test_dragiso(old=None,skip_calcs=False,verbose=False,metals='Cu Fe',**kwargs
             assert isclose(f1,f2,verbose=verbose)
     fname = "drag_iso_fit.txt"
     assert diff(pathlib.Path(old,fname),pathlib.Path(testfolder,fname),verbose=verbose)
-    os.chdir(cwd)
+    os.chdir(new)
 
-def test_drag(old=None,skip_calcs=False,verbose=False,metals='Al Mo Ti Sn',**kwargs):
+def test_drag(old=None,new=cwd,skip_calcs=False,verbose=False,metals='Al Mo Ti Sn',**kwargs):
     '''implements regression tests for anisotropic phonon drag calculations via frontend script dragcoeff_semi_iso.py,
        where folder "old" contains the baseline results; set to "None" to initialize a new baseline.'''
-    testfolder = cwd/"regressiontests"
+    testfolder = new/"regressiontests"
[…]
