### 2026-02-16 a62c16b
don't run legacy test suite on windows (too slow)
also skip most artifacts (not necessary)
