+++ .github/workflows/CI.yml
@@ -39,16 +39,13 @@ jobs:
           python -m pip install --upgrade pip
           pip install ${{ matrix.module-versions }}
           pip install .
-      - name: Run regression test suite
+      - name: Run legacy regression test suite
         run: yes 2>/dev/null | testing/test_pydislocdyn.py . --Ntheta=3 --Nbeta=5 --maxrec=1 --Nchunks=1 --verbose=true --runtests=all | tee test_output.log
         if: runner.os != 'Windows'
-      - name: Run regression test suite windows
-        run: echo 'N' | python testing/test_pydislocdyn.py --Ntheta=3 --Nbeta=5 --maxrec=0 --Nchunks=1 --bccslip=112 --hcpslip=prismatic --verbose=true --runtests=all .
-        if: runner.os == 'Windows'
-      - name: Run unit tests
+      - name: Run tests
         run: yes 2>/dev/null | pytest -s testing/
         if: runner.os != 'Windows'
-      - name: Run unit tests windows
+      - name: Run tests windows
         run: echo 'N' | pytest -s testing/
         if: runner.os == 'Windows'
       - name: Upload main log file as artifact
@@ -57,13 +54,4 @@ jobs:
           name: test-log_${{runner.os}}_${{matrix.python-version}}
           path: test_output.log
         if: runner.os != 'Windows'
-      - name: Upload other log files as artifacts
-        uses: actions/upload-artifact@v4
-        with:
-          name: test-logfiles_${{runner.os}}_${{matrix.python-version}}
-          path: |
-            poly.log
-            dragiso.log
-            drag/dragsemi.log
-            LT_exp/LT.log
 
