### 2026-02-16 a9f36d7
cleanup the new code, support metals=all, skip legacy test script
