+++ testing/test_regression.py
@@ -2,10 +2,8 @@
 # test suite for PyDislocDyn
 # Author: Daniel N. Blaschke
 # Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-# Date: Mar. 6, 2023 - Feb. 15, 2026
-'''This script implements regression testing for PyDislocDyn. Required argument: 'folder' containing old results.
-   (To freshly create a folder to compare to later, run from within an empty folder with argument 'folder' set to '.')
-   For additional options, call this script with '--help'.'''
+# Date: Mar. 6, 2023 - Feb. 16, 2026
+'''This script implements regression testing for PyDislocDyn and is meant to be run with pytest.'''
 import os
 import sys
 import subprocess
@@ -19,8 +17,9 @@ dir_path = pathlib.Path(__file__).resolve().parents[1] / 'pydislocdyn'
 
 from pydislocdyn.metal_data import fcc_metals, bcc_metals, hcp_metals, tetr_metals, \
     ISO_l, c111, all_metals
-from pydislocdyn.utilities import str2bool, isclose, compare_df
+from pydislocdyn.utilities import isclose
 from pydislocdyn import read_2dresults, Ncores, Voigt, strain_poly, writeallinputfiles, readinputfile
+from pydislocdyn.dragcoeff_semi_iso import metal as all_drag_metals
 import numpy as np ## import pydislocdyn first as it will set the openmp thread number
 import sympy as sp
 import pandas as pd
@@ -28,6 +27,8 @@ if Ncores>1:
     from joblib import Parallel, delayed
     
 cwd =pathlib.Path.cwd()
+# all available metals (to be used if user requests 'all' in one of the tests below):
+all_metals = " ".join(all_metals)
 
 ########## define helper functions for the tests below
 
@@ -102,9 +103,7 @@ def expand_slipsystems(metals,bccslip='all',hcpslip='all'):
     '''takes a list of keyword-strings for metals and appends slip system names'''
     if isinstance(metals, str):
         metals = metals.split(" ")
-    elif isinstance(metals, list) and isinstance(metals[0], str):
-        metals = metals
-    else:
+    elif not (isinstance(metals, list) and isinstance(metals[0], str)):
         raise ValueError(f"epxected a string or list of strings but got {metals=}")
     out = []
     if bccslip == 'all':
@@ -131,7 +130,7 @@ def expand_slipsystems(metals,bccslip='all',hcpslip='all'):
 def test_aver(old=None,new=cwd,skip_calcs=False,verbose=False):
     '''implements regression tests for frontend script polycrystal_averaging.py'''
     testfolder = new/"regressiontests"
-    testfolder.mkdir(exist_ok=True)
+    testfolder.mkdir(parents=True,exist_ok=True)
     if old is None:
         old = testfolder
     else:
@@ -152,7 +151,7 @@ def test_dragiso(old=None,new=cwd,skip_calcs=False,verbose=False,metals='Cu Fe',
     '''implements regression tests for isotropic phonon drag calculations via frontend script dragcoeff_iso.py,
        where folder "old" contains the baseline results; set to "None" to initialize a new baseline.'''
     testfolder = new/"regressiontests"
-    testfolder.mkdir(exist_ok=True)
+    testfolder.mkdir(parents=True,exist_ok=True)
     if old is None:
         old = testfolder
     else:
@@ -164,6 +163,15 @@ def test_dragiso(old=None,new=cwd,skip_calcs=False,verbose=False,metals='Cu Fe',
     options['Ncores'] = Ncores
     options.update(kwargs)
     commandargs = convert_options(options)
+    if metals == 'all':
[…]
