+++ .github/workflows/CI.yml
@@ -39,19 +39,19 @@ jobs:
           python -m pip install --upgrade pip
           pip install ${{ matrix.module-versions }}
           pip install .
-      - name: Run legacy regression test suite
-        run: yes 2>/dev/null | testing/test_pydislocdyn.py . --Ntheta=3 --Nbeta=5 --maxrec=1 --Nchunks=1 --verbose=true --runtests=all | tee test_output.log
-        if: runner.os != 'Windows'
       - name: Run tests
         run: yes 2>/dev/null | pytest -s testing/
         if: runner.os != 'Windows'
       - name: Run tests windows
         run: echo 'N' | pytest -s testing/
         if: runner.os == 'Windows'
-      - name: Upload main log file as artifact
+      - name: Upload log files as artifacts
         uses: actions/upload-artifact@v4
         with:
-          name: test-log_${{runner.os}}_${{matrix.python-version}}
-          path: test_output.log
-        if: runner.os != 'Windows'
+          name: test-logfiles_${{runner.os}}_${{matrix.python-version}}
+          path: |
+            regressiontests/poly.log
+            regressiontests/dragiso.log
+            regressiontests/drag/dragsemi.log
+            regressiontests/LT_exp/LT.log
 
