### 2026-02-16 a97e41c
outsource foder-prep (avoids code duplication)
