### 2026-02-16 40a4bd0
split test_misc into 2 tests (test_misc and test_strainpoly)
