+++ testing/test_regression.py
@@ -139,6 +139,15 @@ def prepare_testfolder(old,new,verbose=False):
     print(testfolder)
     print(old)
     return testfolder, old
+
+def prepare_inputfiles(tmpfolder="temp_pydislocdyn"):
+    '''creates the specified temporary folder and writes all input files into that folder'''
+    tmppydislocdyn = pathlib.Path(tmpfolder)
+    tmppydislocdyn.mkdir(exist_ok=True)
+    os.chdir(tmppydislocdyn)
+    writeallinputfiles()
+    os.chdir("..")
+    return tmppydislocdyn
     
 ########## tests:
 
@@ -264,8 +273,8 @@ def test_LT(old=None,new=cwd,skip_calcs=False,verbose=False,metals='Al Mo Ti Sn'
         assert diff(pathlib.Path(old,folder,fname),pathlib.Path(testfolder,folder,fname),verbose=verbose) is True
     os.chdir(new)
 
-def test_acc(old=None,new=cwd,skip_calcs=False,verbose=False,metals='Al Mo Ti Sn',**kwargs):
-    '''implements regression tests for accelerating dislocation solutions,
+def test_acc_screw(old=None,new=cwd,skip_calcs=False,verbose=False,metals='Al Mo Ti Sn',**kwargs):
+    '''implements regression tests for accelerating screw dislocation solutions,
        where folder "old" contains the baseline results; set to "None" to initialize a new baseline.'''
     testfolder, old = prepare_testfolder(old,new,verbose)
     opts = {'bccslip':'all','hcpslip':'all','fastapprox':True} # defaults for this test
@@ -274,24 +283,15 @@ def test_acc(old=None,new=cwd,skip_calcs=False,verbose=False,metals='Al Mo Ti Sn
     if metals=='all':
         metals = all_metals
     metal_acc_screw = []
-    metal_acc_edge = []
     metal_list = expand_slipsystems(metals,bccslip=opts['bccslip'],hcpslip=opts['hcpslip'])
     for X in metal_list:
         if X in fcc_metals or "basal" in X or "prismatic" in X or "pyramidal" in X or X in tetr_metals:
             metal_acc_screw.append(X)
-        if "112" in X or "basal" in X or "prismatic" in X or X in tetr_metals:
-            metal_acc_edge.append(X)
     if not skip_calcs:
         print("running test 'acc' ...")
-        tmppydislocdyn = pathlib.Path("temp_pydislocdyn")
-        tmppydislocdyn.mkdir(exist_ok=True)
-        os.chdir(tmppydislocdyn)
-        writeallinputfiles()
-        os.chdir("..")
+        tmppydislocdyn = prepare_inputfiles()
         uij_acc_screw = {}
         acc_screw = {}
-        uij_acc_edge = {}
-        acc_edge = {}
         ## another dynamic solution:
         ## assume l(t) = adot*t**3/6 ## (i.e. acceleration starts at 0 and increases at rate adot from t>0)
         adot = 6.2e25 ## time-derivative of acceleration, acc is initially zero at time t=0
@@ -303,7 +303,7 @@ def test_acc(old=None,new=cwd,skip_calcs=False,verbose=False,metals='Al Mo Ti Sn
             return eta(x)/(3*x)
         print(f"calculating accelerating screw dislocation fields for {metal_acc_screw}")
         for X in metal_acc_screw:
-            acc_screw[X] = readinputfile(pathlib.Path("temp_pydislocdyn",X),Nphi=50)
+            acc_screw[X] = readinputfile(pathlib.Path(tmppydislocdyn,X),Nphi=50)
[…]
