### 2026-02-16 13328b5
move a helper fct into metal_data
