### 2026-02-17 2f6b8f0
nicer output from tests
