+++ testing/test_regression.py
@@ -105,7 +105,7 @@ def prepare_testfolder(old,new,verbose=False):
     else:
         old = pathlib.Path(new , old)
     if verbose:
-        print(f"{testfolder=}, baseline folder={old}")
+        print(f"\n{testfolder=}, \nbaseline folder={old}\n")
     os.chdir(testfolder)
     return testfolder, old
 
@@ -125,10 +125,10 @@ def test_aver(old=None,new=cwd,skip_calcs=False,verbose=False):
     testfolder, old = prepare_testfolder(old,new,verbose)
     fname = 'averaged_elastic_constants.tex'
     if not skip_calcs:
-        print("running test 'aver' ...")
+        print("\nrunning test 'aver' ...")
         assert runscript('polycrystal_averaging.py',[],'poly.log')==0
-    else: print("skipping test 'aver' as requested")
-    print(f"checking {fname}")
+    else: print("\nskipping test 'aver' as requested")
+    print(f"checking {fname}\n")
     assert diff(pathlib.Path(old,fname),pathlib.Path(testfolder,fname),verbose=verbose)
     os.chdir(new)
 
@@ -151,11 +151,11 @@ def test_dragiso(old=None,new=cwd,skip_calcs=False,verbose=False,metals='Cu Fe',
         metals = metals.strip()
     commandargs.append(f'{metals}')
     if not skip_calcs:
-        print(f"running test 'dragiso' with {commandargs=} ...")
+        print(f"\nrunning test 'dragiso' with {commandargs=} ...")
         assert runscript("dragcoeff_iso.py",commandargs,'dragiso.log')==0
-    else: print("skipping test 'dragiso' as requested")
+    else: print("\nskipping test 'dragiso' as requested")
     metals = metals.split()
-    print(f"\ncomparing dragiso results for: {metals}")
+    print(f"\ncomparing dragiso results for: {metals}\n")
     for X in metals:
         dragname = "drag"
         if options['NT']>1:# and os.access(pathlib.Path(old,f"{dragname}_T_{X}.dat"), os.R_OK):
@@ -189,12 +189,12 @@ def test_drag(old=None,new=cwd,skip_calcs=False,verbose=False,metals='Al Mo Ti S
     drag_folder = pathlib.Path('drag')
     if not skip_calcs:
         drag_folder.mkdir(exist_ok=True)
-        print(f"running test 'drag' with {commandargs=} ...")
+        print(f"\nrunning test 'drag' with {commandargs=} ...")
         os.chdir(pathlib.Path(testfolder,drag_folder))
         assert runscript("dragcoeff_semi_iso.py",commandargs,'dragsemi.log')==0
         os.chdir(testfolder)
-    else: print("skipping test 'drag' as requested")
-    print(f"\ncomparing drag results for: {metals}")
+    else: print("\nskipping test 'drag' as requested")
+    print(f"\ncomparing drag results for: {metals}\n")
     for X in metals:
         dragname = "drag_anis"
         if options['NT']>1 and os.access(pathlib.Path(old,drag_folder,f"{dragname}_T_{X}.dat.xz"), os.R_OK):
@@ -218,25 +218,28 @@ def test_LT(old=None,new=cwd,skip_calcs=False,verbose=False,metals='Al Mo Ti Sn'
     testfolder, old = prepare_testfolder(old,new,verbose)
     opts = {'Nbeta':50,'Ntheta':200,'Ntheta2':4,'Nphi':500,'scale_by_mu':'exp','bccslip':'all','hcpslip':'all'} # defaults for this test
     opts['Ncores'] = Ncores
[…]
