### 2026-02-17 0e95d28
faster defaults for drag calcs, more verbose test output
