+++ testing/test_regression.py
@@ -97,7 +97,7 @@ def convert_options(optiondict):
         out.append(f"--{k}={v}")
     return out
 
-def prepare_testfolder(old,new,verbose=False):
+def prepare_testfolder(old,new,verbose=True):
     '''creates the required folders and chdir into the test folder'''
     testfolder = pathlib.Path(new , "regressiontests")
     testfolder.mkdir(parents=True,exist_ok=True)
@@ -111,7 +111,7 @@ def prepare_testfolder(old,new,verbose=False):
     else:
         old = pathlib.Path(new , old)
     if verbose:
-        print(f"\n{testfolder=}, \nbaseline folder={old}\n")
+        print(f"\n{testfolder=}, \nbaseline folder={old}")
     os.chdir(testfolder)
     return testfolder, old
 
@@ -126,7 +126,7 @@ def prepare_inputfiles(tmpfolder="temp_pydislocdyn"):
     
 ########## tests:
 
-def test_aver(old=baseln,new=cwd,skip_calcs=False,verbose=False):
+def test_aver(old=baseln,new=cwd,skip_calcs=False,verbose=True):
     '''implements regression tests for frontend script polycrystal_averaging.py'''
     testfolder, old = prepare_testfolder(old,new,verbose)
     fname = 'averaged_elastic_constants.tex'
@@ -138,11 +138,11 @@ def test_aver(old=baseln,new=cwd,skip_calcs=False,verbose=False):
     assert diff(pathlib.Path(old,fname),pathlib.Path(testfolder,fname),verbose=verbose)
     os.chdir(new)
 
-def test_dragiso(old=baseln,new=cwd,skip_calcs=False,verbose=False,metals='Cu Fe',**kwargs):
+def test_dragiso(old=baseln,new=cwd,skip_calcs=False,verbose=True,metals='Cu Fe',**kwargs):
     '''implements regression tests for isotropic phonon drag calculations via frontend script dragcoeff_iso.py,
        where folder "old" contains the baseline results; set to "None" to initialize a new baseline.'''
     testfolder, old = prepare_testfolder(old,new,verbose)
-    options = {'Nbeta': 7, 'use_exp_Lame': True, 'phononwind_opts': {'maxrec': 2, 'Nchunks':3, 'target_accuracy': 0.01}, 'NT': 1} # defaults for this test
+    options = {'Nbeta':7, 'phononwind_opts':{'maxrec':2, 'target_accuracy':0.01}} # defaults for this test
     options['Ncores'] = Ncores
     options.update(kwargs)
     commandargs = convert_options(options)
@@ -176,12 +176,12 @@ def test_dragiso(old=baseln,new=cwd,skip_calcs=False,verbose=False,metals='Cu Fe
     assert diff(pathlib.Path(old,fname),pathlib.Path(testfolder,fname),verbose=verbose)
     os.chdir(new)
 
-def test_drag(old=baseln,new=cwd,skip_calcs=False,verbose=False,metals='Al Mo Ti Sn',**kwargs):
+def test_drag(old=baseln,new=cwd,skip_calcs=False,verbose=True,metals='Al Mo Ti Sn',**kwargs):
     '''implements regression tests for anisotropic phonon drag calculations via frontend script dragcoeff_semi_iso.py,
        where folder "old" contains the baseline results; set to "None" to initialize a new baseline.'''
     testfolder, old = prepare_testfolder(old,new,verbose)
-    options = {'Nbeta': 7, 'use_exp_Lame': True, 'phononwind_opts': {'maxrec': 2, 'Nchunks':3, 'target_accuracy': 0.01}, 'NT': 1,
-               'bccslip': 'all', 'hcpslip': 'all', 'Ntheta': 4} # defaults for this test
[…]
