### 2026-02-17 bee8322
fix previous commit (deleted too many default options)
