### 2026-02-18 de30517
Argparse (#2)
.) use argparse for parsing command line options instead of our previous custom parser;
(Note some changed syntax: options '--use_iso', '--allplots', and '--skip_plots' no longer take an argument)
.) port the regression test suite to pytest (testing/test_pydislocdyn.py is deprecated and will be removed soon, use pytest testing/test_regression.py instead)
.) change some defaults for faster dislocation drag calculations
