+++ testing/test_regression.py
@@ -1,0 +1,445 @@
+#!/usr/bin/env python3
+# test suite for PyDislocDyn
+# Author: Daniel N. Blaschke
+# Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
+# Date: Mar. 6, 2023 - Feb. 17, 2026
+'''This script implements regression testing for PyDislocDyn and is meant to be run with pytest.'''
+import os
+import sys
+import subprocess
+import pathlib
+import difflib
+import lzma
+dir_path = str(pathlib.Path(__file__).resolve().parents[1])
+if dir_path not in sys.path:
+    sys.path.append(dir_path)
+dir_path = pathlib.Path(__file__).resolve().parents[1] / 'pydislocdyn'
+
+from pydislocdyn.metal_data import fcc_metals, tetr_metals, ISO_l, c111, all_metals, expand_slipsystems
+from pydislocdyn.utilities import isclose, usefortran
+from pydislocdyn import read_2dresults, Ncores, Voigt, strain_poly, writeallinputfiles, readinputfile
+from pydislocdyn.dragcoeff_semi_iso import metal as all_drag_metals
+import numpy as np ## import pydislocdyn first as it will set the openmp thread number
+import sympy as sp
+import pandas as pd
+if Ncores>1:
+    from joblib import Parallel, delayed
+    
+cwd =pathlib.Path.cwd()
+# all available metals (to be used if user requests 'all' in one of the tests below):
+all_metals = " ".join(all_metals)
+baseln = "baseline" # default folder to check against
+
+########## define helper functions for the tests below
+
+def readfile(fname):
+    '''reads a text file (or xz compressed text file) and returns a list of its lines'''
+    fname = str(fname)
+    if fname[-3:] == '.xz':
+        with lzma.open(fname,"rt") as f1:
+            f1lines = f1.readlines()
+    else:
+        with open(fname,"r", encoding="utf8") as f1:
+            f1lines = f1.readlines()
+    if fname[-4:] == '.tex' or fname[-4:] == '.txt': ## ignore expected changes in LaTeX code
+        f1lines = [x for x in f1lines if 'tabular' not in x and 'table' not in x and 'averages' not in x]
+    return f1lines
+
+def removewhitespace(somestring):
+    '''replaces excessive whitespace with ' ' in the middle of a string, and removes leading/trailing whitespace (space and tab characters only)'''
+    string1 = somestring.strip()
+    string2 = string1.replace("\t"," ").replace("  "," ")
+    while string2!=string1:
+        string1 = string2
+        string2 = string1.replace("\t"," ").replace("  "," ")
+    return string2
+
+def diff(f1,f2,verbose=True):
+    '''Compares two text files'''
+    f1lines = readfile(f1)
+    f1lines = [removewhitespace(i) for i in f1lines]
+    f2lines = readfile(f2)
+    f2lines = [removewhitespace(i) for i in f2lines]
+    thediff = difflib.unified_diff(f1lines, f2lines, fromfile=str(f1),tofile=str(f2),lineterm="",n=0)
+    equal = f1lines==f2lines
+    if verbose and not equal:
+        for line in thediff:
+            print(line.strip("\n"))
+    return equal
+
+def round_list(lst,ndigits=2):
+    '''rounds all floats in a nested list'''
+    if isinstance(lst,float):
+        return float(round(lst,ndigits))
+    if isinstance(lst,list):
+        return [round_list(i,ndigits) for i in lst]
+    return lst
+
[…]
