+++ pydislocdyn/polycrystal_averaging.py
@@ -2,7 +2,7 @@
 # Compute averages of elastic constants for polycrystals
 # Author: Daniel N. Blaschke
 # Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-# Date: Nov. 7, 2017 - Oct. 30, 2025
+# Date: Nov. 7, 2017 - Feb. 13, 2026
 '''This script will compute polycrystal averages of second and third order elastic constants;
    it is not meant to be used as a module. By default, all metals predefined in pydislocdyn.metal_data
    will be taken into account unless the user passes input files (or keywords for some of 
@@ -57,7 +57,7 @@ if __name__ == '__main__':
     metal_kws = metal.copy()
     if len(sys.argv) > 1:
         args = sys.argv[1:]
-        if '--help' in args:
+        if '--help' in args or args[0]=='-h':
             print(f"{__doc__}")
             sys.exit()
         try:
