### 2026-02-18 6e9d7f3
removed obsolete code
- removed the old (legacy) test suite
- removed the custom parser (no longer needed)
- updated the changelog file
