### 2026-02-18 ae8a0dd
updated CI script and changelog-formatting
