+++ .github/workflows/CI.yml
@@ -26,12 +26,12 @@ jobs:
         toolchain:
           - {compiler: gcc, version: 'latest'}
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
       - name: Setup Fortran Compiler
         uses: fortran-lang/setup-fortran@v1
         if: runner.os == 'MacOS'
       - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v4
+        uses: actions/setup-python@v5
         with:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
