### 2026-02-23 3a0fa30
TODO: investigate numpy 2.4-bug, limit to <2.4 for now
