### 2026-02-24 d9f85c4
fix B_of_sigma() for numpy 2.4
