### 2026-02-24 b4ac608
remove option Nchunks>1
to simplify the phononwind code
that old option gave only moderate speedup for the
numba-jit subroutines, but no benefit for the (preferred)
fortran subroutines
