+++ pydislocdyn/phononwind.py
@@ -34,7 +34,7 @@ if Ncores>1:
 
 ### rho x ct^2  = c44, and B is divided by rho^2*ct^4 = c44^2;
 ### it is therefore convenient to divide A3 by c44 as it enters quadratically, and this is a requirement below, i.e. A3 must be rescaled by c44 to be dimensionless!
-def dragcoeff_iso(dij, A3, qBZ, ct, cl, beta, burgers, T, modes='all', Nt=321, Nq1=400, Nphi1=50, Debye_series=False, target_accuracy=5e-3, maxrec=6, accurate_to_digit=1e-5, Nchunks=1, skip_theta=None, skip_theta_val=np.inf, r0cut=-1, name='drag'):
+def dragcoeff_iso(dij, A3, qBZ, ct, cl, beta, burgers, T, modes='all', Nt=321, Nq1=400, Nphi1=50, Debye_series=False, target_accuracy=5e-3, maxrec=6, accurate_to_digit=1e-5, skip_theta=None, skip_theta_val=np.inf, r0cut=-1, name='drag'):
     '''Computes the drag coefficient from phonon wind for an isotropic crystal. Required inputs are the dislocation displacement gradient (times magnitude q and rescaled by the Burgers vector) dij in Fourier space
        (being a 3x3xNthetaxNphi array where theta is the angle parametrizing the dislocation type and phi is the polar angle in Fourier space), the array of shifted 3rd order elastic constants A3 in units of the shear modulus mu,
        the radius of the Brillouin zone qBZ, the transverse and longitudinal sound speeds ct and cl, the velocity beta in units of ct (i.e. beta=v/ct), the magnitude of the Burgers vectors burgers, as well as the temperature T.
@@ -64,28 +64,9 @@ def dragcoeff_iso(dij, A3, qBZ, ct, cl, beta, burgers, T, modes='all', Nt=321, N
         iso=True
     dij=np.moveaxis(dij, -1, 0) ## shape:  (i,j,Ntheta,Nphi) -> (Nphi,i,j,Ntheta)
     A3=np.moveaxis(A3,0,-1) ## shape:  (Ntheta,i,ii,j,jj,k,kk) -> (i,ii,j,jj,k,kk,Ntheta)
-    Nchks = Nchunks
-    ## make sure Nt_total-1 is divisible by 2*Nchunks, and that Nt_current is odd and >=5 (i.e. increase user provided Nt as necessary)
-    Nt_k = int(abs(Nt - 1 - 4*Nchunks)/(2*Nchunks))
-    Nt_total = 4*Nchunks + 2*Nt_k*Nchunks + 1
-    if Nt_total < Nt:
-        Nt_total += 2*Nchunks ## ensure we never reduce resolution
-    
-    def adaptive_t_chunks(dij, A3, qBZ, cs, beta, burgers, T, Nq1=Nq1, Nphi1=Nphi1, Debye_series=Debye_series, beta_long=False, target_accuracy=target_accuracy, \
-                          maxrec=maxrec, accurate_to_digit=accurate_to_digit, skip_theta=skip_theta, r0cut=r0cut, Nt_total=Nt, Nchunks=Nchunks, mode='??'):
-        if Nchunks==1:
-            out = adaptive_t(dij, A3, qBZ, cs, beta, burgers, T, Nq1=Nq1, Nphi1=Nphi1, Debye_series=Debye_series, beta_long=beta_long, target_accuracy=target_accuracy, \
-                             maxrec=maxrec, accurate_to_digit=accurate_to_digit, skip_theta=skip_theta, r0cut=r0cut, Nt=Nt, mode=mode)
-        else:
[…]
