+++ pydislocdyn/subroutines.f90
@@ -2,11 +2,11 @@
 ! run 'python -m numpy.f2py -c subroutines.f90 -m subroutines' to use
 ! Author: Daniel N. Blaschke
 ! Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-! Date: July 23, 2018 - Feb. 22, 2026
+! Date: July 23, 2018 - Feb. 23, 2026
 
 subroutine version(versionnumber)
   integer, intent(out) :: versionnumber
-  versionnumber=20260222
+  versionnumber=20260223
 end subroutine version
 
 module parameters
@@ -654,13 +654,13 @@ END SUBROUTINE computeuij
 
 !!**********************************************************************
 
-SUBROUTINE integratetphi(B,beta,t,phi,updatet,kthchk,Nphi,Nt,Bresult)
+SUBROUTINE integratetphi(B,beta,t,phi,updatet,Nphi,Nt,Bresult)
 ! this is a subroutine of dragcoeff_iso() in phononwind.py
 !-----------------------------------------------------------------------
   use parameters, only : sel
   IMPLICIT NONE
 !-----------------------------------------------------------------------
-  INTEGER, INTENT(IN) :: Nphi, Nt, kthchk
+  INTEGER, INTENT(IN) :: Nphi, Nt
   REAL(KIND=sel), INTENT(IN), DIMENSION(Nt,Nphi)  :: B
   REAL(KIND=sel), INTENT(IN), DIMENSION(Nt)  :: t
   REAL(KIND=sel), INTENT(IN), DIMENSION(Nphi)  :: phi
@@ -682,9 +682,7 @@ SUBROUTINE integratetphi(B,beta,t,phi,updatet,kthchk,Nphi,Nt,Bresult)
     NBtmp = size(Btmp)
     Bt(p) = 0.d0
     if (NBtmp.gt.1) then
-      if ((updatet.eqv..True.).or.(kthchk.eq.0)) then
-        Btmp(1) = 2.d0*Btmp(1)
-      endif
+      Btmp(1) = 2.d0*Btmp(1)
       if (updatet.eqv..True.) then
         Btmp(ubound(Btmp)) = 2.d0*Btmp(ubound(Btmp))
       endif
@@ -699,18 +697,17 @@ END SUBROUTINE integratetphi
 
 !!**********************************************************************
 
-SUBROUTINE integrateqtildephi(B,beta1,qtilde,t,phi,updatet,kthchk,Nchunks,Nphi,Nt,Bresult)
+SUBROUTINE integrateqtildephi(B,beta1,qtilde,t,phi,Nphi,Nt,Bresult)
 ! this is a subroutine of dragcoeff_iso() in phononwind.py
 !-----------------------------------------------------------------------
   use parameters, only : sel
   IMPLICIT NONE
 !-----------------------------------------------------------------------
-  INTEGER, INTENT(IN) :: Nphi, Nt, kthchk, Nchunks
+  INTEGER, INTENT(IN) :: Nphi, Nt
   REAL(KIND=sel), INTENT(IN), DIMENSION(Nt,Nphi)  :: B, t
   REAL(KIND=sel), INTENT(IN), DIMENSION(Nt)  :: qtilde
   REAL(KIND=sel), INTENT(IN), DIMENSION(Nphi)  :: phi
   REAL(KIND=sel), INTENT(IN)  :: beta1
-  LOGICAL, INTENT(IN)  :: updatet
[…]
