### 2026-02-24 2486583
optimize testing defaults in nofortran case
