### 2026-03-19 086edd2
minor: modernize fortran codestyle
