+++ pydislocdyn/subroutines.f90
@@ -53,11 +53,11 @@ output(:,:,:,:,:) = 0.0
 !$ k=i*lent
 !$ call thesum(output(j:k,:,:,:,:),tcosphi(j:k),sqrtsinphi(j:k),tsinphi(j:k),sqrtcosphi(j:k),sqrtt(j:k), &
 !$              qv(j:k,:),delta1,delta2,mag(j:k),A3,phi1,dphi1,lenph1,lent)
-!$ enddo
+!$ end do
 !$OMP END PARALLEL DO
 !$ else
 call thesum(output,tcosphi,sqrtsinphi,tsinphi,sqrtcosphi,sqrtt,qv,delta1,delta2,mag,A3,phi1,dphi1,lenph1,lentph)
-!$ endif
+!$ end if
 end subroutine parathesum
 
 subroutine thesum(output,tcosphi,sqrtsinphi,tsinphi,sqrtcosphi,sqrtt,qv,delta1,delta2,mag,A3,phi1,dphi1,lenph1,lentph)
@@ -209,11 +209,11 @@ SUBROUTINE elbrak(a,b,Cmat,Ntheta,Nphi,AB)
         do l=1,3
           do k=1,3
             AB(:,l,o,i) = AB(:,l,o,i) + a(:,k,i)*Cmat(k,l,o,p,i)*b(:,p,i)
-          enddo
-        enddo
-      enddo
-    enddo
-  enddo
+          end do
+        end do
+      end do
+    end do
+  end do
   !$OMP END PARALLEL DO
   
   RETURN
@@ -238,10 +238,10 @@ SUBROUTINE elbrak1d(a,b,Cmat,Nphi,AB)
       do l=1,3
         do k=1,3
           AB(:,l,o) = AB(:,l,o) + a(:,k)*Cmat(k,l,o,p)*b(:,p)
-        enddo
-      enddo
-    enddo
-  enddo
+        end do
+      end do
+    end do
+  end do
   
   RETURN
 END SUBROUTINE elbrak1d
@@ -366,7 +366,7 @@ SUBROUTINE accscrew_xyintegrand(integrand,x,y,t,xpr,a,b,c,ct,abc,ca,xcomp)
                    - (x-xpr)**2*(tau_min_R)) &
                 - (stepfct2*(1.d0/rpr**4)*((tau2**2*y**2*abc/ct**2 - (x-xpr)*y*(b/(2*c))*rpr**2/(ct*ca**2))/tau_min_R2 &
                    - (x-xpr)**2*tau_min_R2))
-  endif
+  end if
   
   RETURN
 END SUBROUTINE accscrew_xyintegrand
@@ -451,12 +451,12 @@ SUBROUTINE computeEtot(uij, betaj, C2, Cv, phi, Ntheta, Nphi, Wtot)
         do l=1,3
           do k=1,3
             Wdensity(:) = Wdensity(:) + 0.5d0*uij(:,l,k,th)*uij(:,o,p,th)*(C2(k,l,o,p) + betaj*betaj*Cv(k,l,o,p,th))
-          enddo
-        enddo
-      enddo
-    enddo
+          end do
+        end do
+      end do
+    end do
     call trapz(Wdensity,phi,Nphi,Wtot(th))
-  enddo
+  end do
   !$OMP END PARALLEL DO
   
   RETURN
@@ -482,7 +482,7 @@ SUBROUTINE inv(A,invA)
   ! note: hard coding this loop speeds up things by factor 6/27
 !~   do i=1,3; do j=1,3; do k=1,3
 !~     det = det + eps(i,j,k)*A(1,i)*A(2,j)*A(3,k)
-!~   enddo; enddo; enddo
+!~   end do; end do; end do
[…]
