### 2026-03-19 cb10bfa
small code cleanup
make subroutines.version an integer (instead of a fct)
