+++ pydislocdyn/utilities.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 # Author: Daniel N. Blaschke
 # Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-# Date: Nov. 5, 2017 - Feb. 23, 2026
+# Date: Nov. 5, 2017 - Mar. 19, 2026
 '''This module contains various utility functions used by other submodules.'''
 #################################
 import sys
@@ -85,9 +85,10 @@ def ompthreads():
     return 0
 try:
     import pydislocdyn.subroutines as fsub
-    if fsub.version()>=20250914:
+    if  hasattr(fsub.parameters,'version') and fsub.parameters.version>=20260319:
         usefortran = True
         ompthreads = fsub.ompinfo
+        fsub.version = int(fsub.parameters.version)
     else:
         print("Error: the subroutines module is outdated, please re-compile by calling pydislocdy.utilities.compilefortranmodule() and reloading pydislocdyn")
 except ImportError:
