### 2026-03-20 e0aaaa4
update CI.yml
