+++ .github/workflows/CI.yml
@@ -27,12 +27,12 @@ jobs:
         toolchain:
           - {compiler: gcc, version: 'latest'}
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
       - name: Setup Fortran Compiler
         uses: fortran-lang/setup-fortran@v1
         if: runner.os == 'MacOS'
       - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v5
+        uses: actions/setup-python@v6
         with:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
@@ -47,7 +47,7 @@ jobs:
         run: echo 'N' | pytest -s --durations=0 testing/
         if: runner.os == 'Windows'
       - name: Upload log files as artifacts
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@v7
         with:
           name: test-logfiles_${{runner.os}}_${{matrix.python-version}}
           path: |
