### 2026-03-20 167c049
backport updates from master
refactored fortran code
optimized default options for test suite
