+++ pydislocdyn/subroutines.f90
@@ -2,17 +2,13 @@
 ! run 'python -m numpy.f2py -c subroutines.f90 -m subroutines' to use
 ! Author: Daniel N. Blaschke
 ! Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-! Date: July 23, 2018 - Feb. 23, 2026
-
-subroutine version(versionnumber)
-  integer, intent(out) :: versionnumber
-  versionnumber=20260223
-end subroutine version
+! Date: July 23, 2018 - Mar. 19, 2026
 
 module parameters
 implicit none
 integer,parameter :: sel = selected_real_kind(10)
 integer,parameter :: selsm = selected_real_kind(6)  ! some memory-heavy subroutines use lower precision in favor of speed
+integer,parameter :: version = 20260319
 real(kind=sel), parameter :: hbar = 1.0545718d-34       ! reduced Planck constant
 real(kind=sel), parameter :: kB = 1.38064852d-23        ! Boltzmann constant
 real(kind=sel), parameter :: pi = (4.d0*atan(1.d0)) ! pi
@@ -28,164 +24,6 @@ nthreads = 0
 !~ !$   print*, 'type "export OMP_NUM_THREADS=n" before running this prog. to change'
 return
 end subroutine ompinfo
-
-subroutine parathesum(output,tcosphi,sqrtsinphi,tsinphi,sqrtcosphi,sqrtt,qv,delta1,delta2,mag,A3,phi1,dphi1, &
-                      lenp,lent,lenph1,lentph)
-! this wrapper parallelizes thesum() which is a subroutine of dragcoeff_iso_computepoly() in phononwind.py
-!$   Use omp_lib
-use parameters, only : selsm
-implicit none
-
-integer, intent(in) :: lentph, lenph1, lenp, lent
-real(kind=selsm), intent(in), dimension(lenph1) :: phi1, dphi1
-real(kind=selsm), intent(in), dimension(lentph) :: mag, tcosphi, sqrtsinphi, tsinphi, sqrtcosphi, sqrtt
-real(kind=selsm), intent(in), dimension(lentph,3) :: qv
-real(kind=selsm), intent(in), dimension(3,3) :: delta1, delta2
-real(kind=selsm), intent(in), dimension(3,3,3,3,3,3) :: A3
-real(kind=selsm), intent(out), dimension(lentph,3,3,3,3) :: output
-!$ integer :: i,j,k, nthreads
-output(:,:,:,:,:) = 0.0
-!$ call ompinfo(nthreads)
-!$ if (nthreads .ge. 2) then
-!$OMP PARALLEL DO default(shared), private(i,j,k)
-!$ do i=1,lenp
-!$ j=(i-1)*lent+1
-!$ k=i*lent
-!$ call thesum(output(j:k,:,:,:,:),tcosphi(j:k),sqrtsinphi(j:k),tsinphi(j:k),sqrtcosphi(j:k),sqrtt(j:k), &
-!$              qv(j:k,:),delta1,delta2,mag(j:k),A3,phi1,dphi1,lenph1,lent)
-!$ enddo
-!$OMP END PARALLEL DO
-!$ else
-call thesum(output,tcosphi,sqrtsinphi,tsinphi,sqrtcosphi,sqrtt,qv,delta1,delta2,mag,A3,phi1,dphi1,lenph1,lentph)
-!$ endif
-end subroutine parathesum
-
[…]
