+++ changelog.md
@@ -6,7 +6,7 @@ Features and improvements:
 
  - use argparse for parsing command line options instead of our previous custom parser;
 (Note some changed syntax: options `--use_iso`, `--allplots`, and `--skip_plots` no longer take an argument)
- - ported the test suite to pytest
+ - ported the test suite to pytest and optimized its defaults for faster completion
  - changed some defaults for faster dislocation drag calculations
 
 Fix:
@@ -17,6 +17,7 @@ Other:
 
  - use pathlib for paths
  - removed deprecated keyword Nchunks from phonondrag()
+ - refactored the Fortran code for better readability and `version` is now an integer instead of a fct
 
 ## 1.3.4 (2025-11-18)
 
