### 2026-03-25 4be9a57
add more unit tests
for (some of) the fortran subroutines
tests are implemented in test_unit.py (for pytest after f2py compilation)
as well as in runtests.f90 (as a standalone option)
