+++ pydislocdyn/runtests.f90
@@ -1,0 +1,98 @@
+! compile with:
+! gfortran pydislocdyn/subroutines.f90 pydislocdyn/runtests.f90 -o runtests.x -Wall -pedantic -Wextra -fopenmp -lgomp
+! or similar
+module checks
+  contains
+  subroutine testequal(A,B,n,m,string,tolerance,count_pass,count_fail)
+  !  Check if two NxM matrices are equal (with some tolerance), and print 'test <string> PASSED/FAILED'
+    implicit none
+    integer, intent(in)  :: n, m
+    real(kind=8), intent(in)  :: tolerance
+    real(kind=8), dimension(n,m), intent(in)  :: A, B
+    character(*), intent(in) :: string
+    integer :: count_pass,count_fail
+    logical :: equal
+    
+    equal = all(abs(A-B) <= tolerance)
+    if (equal) then
+      count_pass = count_pass+1
+      print*,"test " // string//": "//char(9)//"PASSED"
+    else
+      count_fail = count_fail+1
+      print*,"test " // string//": "//char(9)//"FAILED"
+    endif
+    
+    return
+  end subroutine testequal
+
+  SUBROUTINE testequalarray(A,B,n,string,tolerance,count_pass,count_fail)
+  !  Check if two arrays are equal (with some tolerance), and print 'test <string> PASSED/FAILED'
+    implicit none
+    integer, intent(in)  :: n
+    real(kind=8), intent(in)  :: tolerance
+    real(kind=8), dimension(n), intent(in)  :: A, B
+    character(*), intent(in) :: string
+    integer :: count_pass,count_fail
+    logical :: equal
+    
+    equal = all(abs(A-B) <= tolerance)
+    if (equal) then
+      count_pass = count_pass+1
+      print*,"test " // string//": "//char(9)//"PASSED"
+    else
+      count_fail = count_fail+1
+      print*,"test " // string//": "//char(9)//"FAILED"
+    endif
+    
+    return
+  end subroutine testequalarray
+
+  subroutine testzero(A,string,tolerance,count_pass,count_fail)
+  ! check if A=0 (with some tolerance), and print 'test <string> PASSED/FAILED'
+    implicit none
+    real(kind=8), intent(in)  :: tolerance, A
+    character(*), intent(in) :: string
+    integer :: count_pass,count_fail
+    call testequalarray((/A/),(/0.d0/),1,string,tolerance,count_pass,count_fail)
+    return
+  end subroutine testzero
+end module checks
+
+program runtests
+  use parameters
+  use phononwind_subroutines
+  use checks
+  implicit none
+  
+  real(kind=8) :: tmpintegral, array1(5),array2(5), start_time, finish_time
+  real(kind=8), dimension(3,3) :: A, B, one=reshape((/1.d0,0.d0,0.d0,0.d0,1.d0,0.d0,0.d0,0.d0,1.d0/),(/3,3/))
+  real(kind=8), allocatable, dimension(:) :: x, func, integral
+  integer :: resol, count_fail=0, count_pass=0
+  resol = 10000
+  allocate(x(resol), func(resol), integral(resol))
+  call cpu_time(start_time)
+  
+  ! test linspace
+  call linspace(0.2d0,1.d0,5,array1)
+  array2 = (/0.2d0,0.4d0,0.6d0,0.8d0,1.d0/)
+  call testequalarray(array1,array2,5,"linspace",1.d-15,count_pass,count_fail)
+  
+  ! test cumtrapz / trapz
+  call linspace(1.d-15, 100.d0, resol, x)
+  func(:) = sin(pi*x(:)/25.d0) + x(:)**3 / (exp(x(:))-1.d0)
[…]
