### 2026-03-25 b8e7a28
check for openmp, make thesum() private, others public
