+++ pydislocdyn/subroutines.f90
@@ -2,13 +2,13 @@
 ! run 'python -m numpy.f2py -c subroutines.f90 -m subroutines' to use
 ! Author: Daniel N. Blaschke
 ! Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-! Date: July 23, 2018 - Mar. 19, 2026
+! Date: July 23, 2018 - Mar. 25, 2026
 
 module parameters
 implicit none
 integer,parameter :: sel = selected_real_kind(10)
 integer,parameter :: selsm = selected_real_kind(6)  ! some memory-heavy subroutines use lower precision in favor of speed
-integer,parameter :: version = 20260319
+integer,parameter :: version = 20260325
 real(kind=sel), parameter :: hbar = 1.0545718d-34       ! reduced Planck constant
 real(kind=sel), parameter :: kB = 1.38064852d-23        ! Boltzmann constant
 real(kind=sel), parameter :: pi = (4.d0*atan(1.d0)) ! pi
@@ -467,6 +467,8 @@ END SUBROUTINE computeuij
 module phononwind_subroutines
   ! his module contains subroutines for phononwind_xx() and phononwind_xy()
   implicit none
+  private :: thesum
+  public :: phonondistri, parathesum, dragintegrand, integratetphi, integrateqtildephi
   contains
     SUBROUTINE phonondistri(prefac,T,c1qBZ,c2qBZ,q1,q1h4,OneMinBtqcosph1,lenq1,lent,lenphi,distri)
       use parameters, only : sel, hbar, kb
