### 2026-03-25 e607c02
more unit tests
