### 2026-03-25 9d5c18e
bump version requirements
