+++ pydislocdyn/utilities.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 # Author: Daniel N. Blaschke
 # Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-# Date: Nov. 5, 2017 - Mar. 19, 2026
+# Date: Nov. 5, 2017 - Mar. 25, 2026
 '''This module contains various utility functions used by other submodules.'''
 #################################
 import sys
@@ -115,13 +115,13 @@ def compilefortranmodule(buildopts='',clean=False):
     compilerflags = '' ## when in doubt, build without OpenMP support
     if sys.version_info[:2]<=(3,11):
         import setuptools
-        if setuptools.__version__ < '70' and np.__version__<'2.0':
+        if setuptools.__version__ < '70' and np.__version__<'1.26':
             if shutil.which('gfortran'):
                 compilerflags = '--f90flags=-fopenmp -lgomp' # flags specific to gfortran, require f2py with (old) distutils backend
         elif np.__version__>='1.26':
             compilerflags = '--dep=openmp --backend=meson' # requires meson to be installed
         else:
-            raise Exception("I need either (setuptools <= 69) or (numpy>=1.26 and meson) to compile.")
+            raise Exception("I need either (setuptools <= 69 and numpy 1.25) or (numpy>=1.26 and meson) to compile.")
     elif sys.version_info[:2]>=(3,12):
         compilerflags = '--dep=openmp' # requires f2py with new meson backend (default in numpy>=2.0 and python>=3.12)
     if buildopts != '':
