+++ changelog.md
@@ -18,6 +18,8 @@ Other:
  - use pathlib for paths
  - removed deprecated keyword Nchunks from phonondrag()
  - refactored the Fortran code for better readability and `version` is now an integer instead of a fct
+ - numpy>=1.25 is now required (older versions cannot handle `private` statements in fortran modules);
+   this has implications for other version requirements (e.g. numba)
 
 ## 1.3.4 (2025-11-18)
 
