+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=61.0", "wheel", "numpy>=1.19"]
+requires = ["setuptools>=61.0", "wheel", "numpy>=1.25"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
@@ -23,16 +23,16 @@ classifiers = [
     "Private :: Do Not Upload",
 ]
 dependencies = [
-  "numpy>=1.19",
+  "numpy>=1.25",
   "scipy>=1.9",
   "sympy>=1.9",
-  "matplotlib>=3.3",
+  "matplotlib>=3.4",
   "pandas>=1.3",
   "jinja2>=2.10",
 ]
 [project.optional-dependencies]
 fast = [
-  "numba>=0.53",
+  "numba>=0.58",
   "joblib>=1.1",
   "threadpoolctl"
 ]
