### 2026-03-25 e9375f7
fix min scipy version
