### 2026-03-25 20e0fee
revert the old-f2py-breaking change
