+++ .github/workflows/CI.yml
@@ -21,7 +21,7 @@ jobs:
             python-version: '3.9'
         include:
           - python-version: '3.9'
-            module-versions: "numpy==1.25 scipy==1.10 sympy==1.9 matplotlib==3.4 pandas==1.3 jinja2 joblib==1.1 threadpoolctl gmpy2 numba==0.58 meson ninja pytest setuptools==69"
+            module-versions: "numpy==1.21 scipy==1.9 sympy==1.9 matplotlib==3.3 pandas==1.3 jinja2 joblib==1.1 threadpoolctl gmpy2 numba==0.53 meson ninja pytest setuptools==69"
           - python-version: '3.14'
             module-versions: "numpy scipy sympy matplotlib pandas jinja2 joblib threadpoolctl gmpy2 numba meson ninja pytest"
         toolchain:
