### 2026-03-26 107bddf
debug / check the new tests, skip many old ones for now
