### 2026-03-26 66d0ccc
try getting f2py to work with teh windows runner
