+++ .github/workflows/CI.yml
@@ -30,7 +30,9 @@ jobs:
       - uses: actions/checkout@v6
       - name: Setup Fortran Compiler
         uses: fortran-lang/setup-fortran@v1
-        if: runner.os == 'MacOS'
+        if: runner.os == 'MacOS' || runner.os == 'Windows'
+        with:
+          compiler: gcc
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v6
         with:
@@ -44,7 +46,7 @@ jobs:
         run: yes 2>/dev/null | pytest -s --durations=0 testing/
         if: runner.os != 'Windows'
       - name: Run tests windows
-        run: echo 'N' | pytest -s --durations=0 testing/
+        run: echo 'Y' | pytest -s --durations=0 testing/
         if: runner.os == 'Windows'
       - name: Upload log files as artifacts
         uses: actions/upload-artifact@v7
