### 2026-03-26 9e878fe
try flang on windows and macos
