### 2026-03-26 e37a593
try again with gfortran on windows
