+++ .github/workflows/CI.yml
@@ -19,6 +19,8 @@ jobs:
         exclude:
           - os: macos-latest
             python-version: '3.9'
+          - os: windows-latest
+            python-version: '3.14'
         include:
           - python-version: '3.9'
             module-versions: "numpy==1.21 scipy==1.9 sympy==1.9 matplotlib==3.3 pandas==1.3 jinja2 joblib==1.1 threadpoolctl gmpy2 numba==0.53 meson ninja pytest setuptools==69"
