### 2026-03-26 fc40301
try with intel compiler on win and mac
