### 2026-03-29 2df4fff
expand last unit test
to include also isotropic inputfiles
