+++ testing/test_unit.py
@@ -2,7 +2,7 @@
 # test suite for PyDislocDyn
 # Author: Daniel N. Blaschke
 # Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-# Date: Mar. 6, 2023 - Mar. 26, 2026
+# Date: Mar. 6, 2023 - Mar. 29, 2026
 '''This script implements several unit tests for PyDislocyn meant to be called by pytest.'''
 import os
 import sys
@@ -36,6 +36,7 @@ def initialize_dislocs(metal_list=None,Ntheta=2):
     '''writes all input files to subfolder "temp_pydislocdyn", then reads a subset defined by the keywords in "metal_list",
        and returns a dictionary of the resulting Dislocation class instances.'''
     os.chdir(tmppydislocdyn)
+    pydis.writeallinputfiles(iso=True)
     pydis.writeallinputfiles()
     Y = {}
     if metal_list is None:
@@ -145,8 +146,11 @@ def test_disloc_props(metal_list=None,Ntheta=2):
     for X in Y:
         Y[X].alignC2()
         Y[X].computevcrit()
-        num_edge = sorted(Y[X].vcrit_barnett[0,-1])[:2]
-        num_screw = sorted(Y[X].vcrit_barnett[0,0])[:2]
+        if Y[X].sym=='iso':
+            num_edge = num_screw = Y[X].ct ## Barnett routine is not called in the isotropic case
+        else:
+            num_edge = sorted(Y[X].vcrit_barnett[0,-1])[:2]
+            num_screw = sorted(Y[X].vcrit_barnett[0,0])[:2]
         if not np.any(np.isclose(num_edge,Y[X].vcrit_edge,rtol=1e-02)):
             print(f"Warning: numerical accuracy of vcrit_edge for {X} may be less than 1%")
         ## need high tolerance in assert statements since numerical barnett scheme is inaccurate in highly symmetric cases
