### 2026-03-31 298986e
add fortran version of strohgeometry and test against python version
