### 2026-03-31 091bccd
bump requirements to fix fortran compilation with old versions
