+++ pydislocdyn/utilities.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 # Author: Daniel N. Blaschke
 # Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-# Date: Nov. 5, 2017 - Mar. 25, 2026
+# Date: Nov. 5, 2017 - Mar. 27, 2026
 '''This module contains various utility functions used by other submodules.'''
 #################################
 import sys
@@ -85,7 +85,7 @@ def ompthreads():
     return 0
 try:
     import pydislocdyn.subroutines as fsub
-    if  hasattr(fsub.parameters,'version') and fsub.parameters.version>=20260319:
+    if  hasattr(fsub.parameters,'version') and fsub.parameters.version>=20260330:
         usefortran = True
         ompthreads = fsub.ompinfo
         fsub.version = int(fsub.parameters.version)
@@ -112,18 +112,7 @@ def compilefortranmodule(buildopts='',clean=False):
        Keyword 'buildopts' may be used to pass additional options to f2py.
        To delete files created by this function, set "clean"=True.'''
     cwd =pathlib.Path.cwd()
-    compilerflags = '' ## when in doubt, build without OpenMP support
-    if sys.version_info[:2]<=(3,11):
-        import setuptools
-        if setuptools.__version__ < '70' and np.__version__<'1.26':
-            if shutil.which('gfortran'):
-                compilerflags = '--f90flags=-fopenmp -lgomp' # flags specific to gfortran, require f2py with (old) distutils backend
-        elif np.__version__>='1.26':
-            compilerflags = '--dep=openmp --backend=meson' # requires meson to be installed
-        else:
-            raise Exception("I need either (setuptools <= 69 and numpy 1.25) or (numpy>=1.26 and meson) to compile.")
-    elif sys.version_info[:2]>=(3,12):
-        compilerflags = '--dep=openmp' # requires f2py with new meson backend (default in numpy>=2.0 and python>=3.12)
+    compilerflags = '--dep=openmp'
     if buildopts != '':
         compilerflags += f" {buildopts}"
     os.chdir(pathlib.Path(__file__).parent)
@@ -142,10 +131,8 @@ def compilefortranmodule(buildopts='',clean=False):
         with open(fname,"w", encoding="utf8") as f1:
             f1.write(f"{error}")
         print(f"\nERROR: compilefortranmodule() failed using {compilerflags=}")
-        print("make sure a Fortran compiler that is supported by numpy.f2py is installed")
-        if '--dep' in compilerflags:
-            print("as well as meson;")
-            print("additional options (if necessary), such as e.g. '--build-dir', may be passed via my 'buildopts' keyword.")
+        print("make sure a Fortran compiler that is supported by numpy.f2py as well as meson are installed")
+        print("additional options (if necessary), such as e.g. '--build-dir', may be passed via my 'buildopts' keyword.")
     elif os.path.isfile(fname):
         os.remove(fname)
     os.chdir(cwd)
