+++ readme.md
@@ -33,22 +33,21 @@ The LANL development team asks that any forks or derivative works include approp
 ## Requirements
 
 * Python >=3.9,</br>
-* [numpy](https://numpy.org/doc/stable/user/) >=1.19,</br>
-* [scipy](https://docs.scipy.org/doc/scipy/reference/) >=1.9,</br>
-* [sympy](https://www.sympy.org) >=1.9,</br>
-* [matplotlib](https://matplotlib.org/) >=3.3</br>
-* [pandas](https://pandas.pydata.org/) >=1.3 (and Jinja2)</br>
+* [numpy](https://numpy.org/doc/stable/user/) >=1.26,</br>
+* [scipy](https://docs.scipy.org/doc/scipy/reference/) >=1.10,</br>
+* [sympy](https://www.sympy.org) >=1.12,</br>
+* [matplotlib](https://matplotlib.org/) >=3.8</br>
+* [pandas](https://pandas.pydata.org/) >=1.5 (and Jinja2)</br>
 * [mpmath](https://mpmath.org/) (recommended optional dependency of mpmath: gmpy)
 
 ### Optional (but recommended):
 
-* a Fortran 90 compiler
+* a Fortran 90 compiler and meson
 to employ the alternative faster Fortran implementations of some subroutines via [f2py](https://docs.scipy.org/doc/numpy/f2py/);</br>
 A helper function, pydislocdyn.utilities.compilefortranmodule(), is included to automate compilation of the Fortran submodule and to ensure it is placed in the correct location.
-</br>Note: building with numpy <1.26 requires setuptools<70; building with numpy >=1.26 requires meson.</br>
 * [joblib](https://joblib.readthedocs.io) >=1.1 (for parallelization),</br>
 * [threadpoolctl](https://github.com/joblib/threadpoolctl) (for automatically adjusting the number of OpenMP threads in the Fortran subroutines and numpy to avoid overcommitting if joblib is used)
-* [numba](https://numba.pydata.org/) >=0.53 (for speedup via just-in-time compilation of some subroutines, although the Fortran subroutines are faster and thus preferred),</br>
+* [numba](https://numba.pydata.org/) >=0.58.1 (for speedup via just-in-time compilation of some subroutines, although the Fortran subroutines are faster and thus preferred),</br>
 * [jupyter](https://jupyter.org/) to view and run the examples notebook
 * a recent version of LaTeX to build the manual (LA-UR-22-28074)
 
