### 2026-03-31 f9e572e
fix unit tests for numba fallback
