+++ pydislocdyn/subroutines.f90
@@ -2,13 +2,13 @@
 ! run 'python -m numpy.f2py -c subroutines.f90 -m subroutines' to use
 ! Author: Daniel N. Blaschke
 ! Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-! Date: July 23, 2018 - Mar. 25, 2026
+! Date: July 23, 2018 - Mar. 30, 2026
 
 module parameters
 implicit none
 integer,parameter :: sel = selected_real_kind(10)
 integer,parameter :: selsm = selected_real_kind(6)  ! some memory-heavy subroutines use lower precision in favor of speed
-integer,parameter :: version = 20260325
+integer,parameter :: version = 20260330
 real(kind=sel), parameter :: hbar = 1.0545718d-34       ! reduced Planck constant
 real(kind=sel), parameter :: kB = 1.38064852d-23        ! Boltzmann constant
 real(kind=sel), parameter :: pi = (4.d0*atan(1.d0)) ! pi
@@ -24,6 +24,119 @@ nthreads = 0
 !~ !$   print*, 'type "export OMP_NUM_THREADS=n" before running this prog. to change'
 return
 end subroutine ompinfo
+
+module elastic_constants
+  implicit none
+  integer, parameter :: VoigtIndices(6)= (/1,5,9,6,3,2/), UnVoigtIndices(9)= (/1,6,5,6,2,4,5,4,3/)
+  private VoigtIndices, UnVoigtIndices ! f2py-incompatibilities prevent us from making more stuff private
+  public voigt, unvoigt
+  interface voigt
+    module procedure vgt_two, vgt_four, vgt_six
+  end interface voigt
+  interface unvoigt
+    module procedure unvgt_one, unvgt_two, unvgt_three
+  end interface unvoigt
+  contains
+    subroutine elasticC2(cij,sym,vc)
+    ! takes a list of indep. elastic constants and returns the 2nd order tensor in Voigt notation
+      use parameters, only : sel
+      real(sel), intent(in) :: cij(:)
+      character(*), intent(in) :: sym
+      real(sel), intent(out) :: vc(6,6)
+      real(sel) :: x, xij(21)
+      integer :: i,j,n,ii,jj
+      xij = 0.d0
+      vc = 0.d0
+      n = size(cij)
+      select case (trim(sym))
+        case ("iso")
+          if (n/=2) then
+            print*,"error: expected length 2"
+            return
+          end if
+          x = cij(1)+2.d0*cij(2) ! user passes (c12,c44) in this case, so x=c11
+          xij(1) = x; xij(7) = x; xij(12) = x
+          xij(2) = cij(1); xij(3) = cij(1); xij(8) = cij(1)
+          xij(16) = cij(2); xij(19) = cij(2); xij(21) = cij(2)
+        case ("cubic")
+          if (n/=3) then
+            print*,"error: expected length 3"
+            return
+          end if
+          xij(1) = cij(1); xij(7) = cij(1); xij(12) = cij(1)
+          xij(2) = cij(2); xij(3) = cij(2); xij(8) = cij(2)
+          xij(16) = cij(3); xij(19) = cij(3); xij(21) = cij(3)
+        case ("hcp")
+          if (n/=5) then
+            print*,"error: expected length 5"
+            return
+          end if
+          x = 0.5d0*(cij(1)-cij(2))
+          xij(1) = cij(1); xij(7) = cij(1); xij(12) = cij(4)
+          xij(2) = cij(2); xij(3) = cij(3); xij(8) = cij(3)
[…]

[…]
