### 2026-03-31 e77f93d
compile/run fortran testsuite as well
