### 2026-03-31 300c9f5
skip intel on linux
