+++ .github/workflows/CI.yml
@@ -22,6 +22,8 @@ jobs:
         exclude:
           - os: macos-latest
             toolchain: {compiler: intel}
+          - os: ubuntu-latest
+            toolchain: {compiler: intel}
           - os: windows-latest
             toolchain: {compiler: gcc}
           - os: macos-latest
@@ -52,7 +54,7 @@ jobs:
           pip install .
       - name: Run tests
         run: |
-          gfortran pydislocdyn/subroutines.f90 pydislocdyn/runtests.f90 -o runtests.x -fopenmp -lgomp
+          gfortran pydislocdyn/subroutines.f90 pydislocdyn/runtests.f90 -o runtests.x -fopenmp -lgomp -Wall -pedantic -Wextra
           ./runtests.x
           yes 2>/dev/null | pytest -s --durations=0 testing/
         if: runner.os != 'Windows'
