### 2026-03-31 0c9f3f8
include fortran testsuite in CI
