### 2026-03-31 36ba1ae
move elasticconstants into its own file
