### 2026-03-31 85eed4c
move to hatchling, include Makefile for runtests.x
