### 2026-03-31 70cf1c5
move elasticconstants into its own file
and include a Makefile to simplify building the fortran testsuite
