### 2026-04-07 6057a68
fix a typo in a doc-string and update the changelog
