### 2026-04-07 fa705ab
add another small test
