### 2026-04-07 294a115
don't rely on default being voigt=false (may change)
