### 2026-04-07 750029a
fix typo in docstring and dont rely on old defaults
