### 2026-04-07 9a3613b
fortran: complete elasticC2, implement voigt-reuss-hill averaging
