+++ testing/test_unit.py
@@ -2,7 +2,7 @@
 # test suite for PyDislocDyn
 # Author: Daniel N. Blaschke
 # Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-# Date: Mar. 6, 2023 - Mar. 30, 2026
+# Date: Mar. 6, 2023 - Apr. 7, 2026
 '''This script implements several unit tests for PyDislocyn meant to be called by pytest.'''
 import os
 import sys
@@ -177,6 +177,11 @@ def test_disloc_props(metal_list=None,Ntheta=2):
             if X in pydis.metal_data.fcc_metals: 
                 vlim_edge = np.sqrt(min(Y[X].cp,Y[X].c44)/Y[X].rho)
                 assert np.isclose(Y[X].vcrit_edge,vlim_edge)
+    if pydis.usefortran:
+        ## check that fortran implementation of averaging schemes matches the python implementations
+        for X in Y:
+            if X in ('Cu','Tibasal','Sn','Mo'): # check only one rep. for each symmetry (Mo is isotropic in contrast to Mo110 etc.)
+                assert np.allclose(np.array(Y[X].compute_Lame(scheme='hill',roundto=0)[:2]),np.array(pydis.subroutines.elastic_constants.hillaverage(Y[X].C2)))
 
 def test_fortransubroutines():
     '''tests some of the fortran code, if it is available'''
@@ -194,11 +199,16 @@ def test_fortransubroutines():
     x = np.random.rand(10)
     assert np.allclose(trapezoid(f,x),pydis.subroutines.trapz(f,x))
     assert np.allclose(cumulative_trapezoid(f,x,initial=0),pydis.subroutines.cumtrapz(f,x))
-    ## check that fortran implementations of (un)voigt and elasticC2 give the same results as the python version
-    xtric = np.linspace(1,21,21)
+    ## check that fortran implementations of elasticC2 and (un)voigt give the same results as the python version
+    xtric = np.random.rand(21)
     assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:2],sym='iso')==pydis.elasticC2(c12=xtric[0],c44=xtric[1],voigt=True))
     assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:3],sym='cubic')==pydis.elasticC2(c11=xtric[0],c12=xtric[1],c44=xtric[2],voigt=True))
     assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:5],sym='hcp')==pydis.elasticC2(c11=xtric[0],c12=xtric[1],c13=xtric[2],c33=xtric[3],c44=xtric[4],voigt=True))
+    assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:6],sym='tetr')==pydis.elasticC2(c11=xtric[0],c12=xtric[1],c13=xtric[2],c33=xtric[3],c44=xtric[4],c66=xtric[5],voigt=True))
+    assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:6],sym='trig')==pydis.elasticC2(cij=xtric[:6],voigt=True))
+    assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:7],sym='tetr2')==pydis.elasticC2(cij=xtric[:7],voigt=True))
+    assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:9],sym='ortho')==pydis.elasticC2(cij=xtric[:9],voigt=True))
+    assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:13],sym='mono')==pydis.elasticC2(cij=xtric[:13],voigt=True))
[…]
