+++ pydislocdyn/elasticconstants.f90
@@ -1,11 +1,11 @@
 ! Author: Daniel N. Blaschke
 ! Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-! Date: Mar. 30, 2026 - Mar. 31, 2026
+! Date: Mar. 30, 2026 - Apr. 7, 2026
 module elastic_constants
   implicit none
   integer, parameter :: VoigtIndices(6)= (/1,5,9,6,3,2/), UnVoigtIndices(9)= (/1,6,5,6,2,4,5,4,3/)
   private VoigtIndices, UnVoigtIndices ! f2py-incompatibilities prevent us from making more stuff private
-  public voigt, unvoigt
+  public voigt, unvoigt, elasticC2, voigtaverage, reussaverage, hillaverage
   interface voigt
     module procedure vgt_two, vgt_four, vgt_six
   end interface voigt
@@ -47,10 +47,43 @@ module elastic_constants
             print*,"error: expected length 5"
             return
           end if
-          x = 0.5d0*(cij(1)-cij(2))
-          xij(1) = cij(1); xij(7) = cij(1); xij(12) = cij(4)
-          xij(2) = cij(2); xij(3) = cij(3); xij(8) = cij(3)
-          xij(16) = cij(5); xij(19) = cij(5); xij(21) = x
+          xij(:3) = cij(:3); xij(7) = cij(1); xij(8) = cij(3); xij(12) = cij(4)
+          xij(16) = cij(5); xij(19) = cij(5); xij(21) = 0.5d0*(cij(1)-cij(2))
+        case ("tetr")
+          if (n/=6) then
+            print*,"error: expected length 6"
+            return
+          end if
+          xij(:3) = cij(:3); xij(7) = cij(1); xij(8) = cij(3); xij(12) = cij(4)
+          xij(16) = cij(5); xij(19) = cij(5); xij(21) = cij(6)
+        case ("trig")
+          if (n/=6) then
+            print*,"error: expected length 6"
+            return
+          end if
+          xij(:4) = cij(:4); xij(7) = cij(1); xij(8) = cij(3); xij(9) = -cij(4); xij(12) = cij(5)
+          xij(16) = cij(6); xij(19) = cij(6); xij(20) = cij(4); xij(21) = 0.5d0*(cij(1)-cij(2))
+        case ("tetr2")
+          if (n/=7) then
+            print*,"error: expected length 7"
+            return
+          end if
+          xij(:3) = cij(:3); xij(6) = cij(4); xij(7) = cij(1); xij(8) = cij(3); xij(11) = -cij(4); xij(12) = cij(5)
+          xij(16) = cij(6); xij(19) = cij(6); xij(21) = cij(7)
+        case ("ortho")
+          if (n/=9) then
+            print*,"error: expected length 9"
+            return
+          end if
+          xij(:3) = cij(:3); xij(7) = cij(4); xij(8) = cij(5); xij(12) = cij(6)
[…]
