### 2026-04-08 bc3a5bc
fortran: impl. kroener average and partially elasticC3
