+++ pydislocdyn/elasticconstants.f90
@@ -34,7 +34,7 @@ module elastic_constants
           xij(1) = x; xij(7) = x; xij(12) = x
           xij(2) = cij(1); xij(3) = cij(1); xij(8) = cij(1)
           xij(16) = cij(2); xij(19) = cij(2); xij(21) = cij(2)
-        case ("cubic")
+        case ("cubic", "fcc", "bcc")
           if (n/=3) then
             print*,"error: expected length 3"
             return
@@ -70,7 +70,7 @@ module elastic_constants
           end if
           xij(:3) = cij(:3); xij(6) = cij(4); xij(7) = cij(1); xij(8) = cij(3); xij(11) = -cij(4); xij(12) = cij(5)
           xij(16) = cij(6); xij(19) = cij(6); xij(21) = cij(7)
-        case ("ortho")
+        case ("orth","ortho")
           if (n/=9) then
             print*,"error: expected length 9"
             return
@@ -91,17 +91,78 @@ module elastic_constants
           end if
           xij = cij
         case default
-          print*,"Error: expected one of these keywords for sym: 'iso', 'cubic', 'hcp', 'tetr', 'trig', 'tetr2', 'ortho', 'mono', 'tric'."
+          print*,"Error: keyword sym must be one of 'iso', 'cubic', 'hcp', 'tetr', 'trig', 'tetr2', 'orth', 'mono', 'tric'."
           return
       end select
       do i=1,6
         do j=1,6
           ii = min(i,j)
           jj = max(i,j)
-          vc(i,j) = xij((jj-ii+1)+int((ii-1)*(14-ii)/2))!xij(int((ii-1)*ii/2)+jj)
+          vc(i,j) = xij((jj-ii+1)+int((ii-1)*(14-ii)/2))
         end do
       end do
     end subroutine elasticC2
+    !-------------------------
+    subroutine elasticC3(cijk,sym,vc)
+    ! takes a list of indep. elastic constants and returns the 2nd order tensor in Voigt notation
+      use parameters, only : sel
+      real(sel), intent(in) :: cijk(:)
+      character(*), intent(in) :: sym
+      real(sel), intent(out) :: vc(6,6,6)
+      real(sel) :: x, xijk(56)
+      integer :: i,j,k,n,ii,jj,kk,iind(6)
+      iind = (/0,15,25,31,34,35/)
+      xijk = 0.d0
+      vc = 0.d0
+      n = size(cijk)
+      select case (trim(sym))
+        case ("iso")
+          if (n/=3) then
+            print*,"error: expected length 3"
+            return
+          end if
[…]
