### 2026-04-08 ec20ba7
fortran: complete elasticC3, continue disloc. class
