+++ testing/test_unit.py
@@ -202,20 +202,28 @@ def test_fortransubroutines():
     assert np.allclose(trapezoid(f,x),pydis.subroutines.trapz(f,x))
     assert np.allclose(cumulative_trapezoid(f,x,initial=0),pydis.subroutines.cumtrapz(f,x))
     ## check that fortran implementations of elasticC2, elasticC3, and (un)voigt give the same results as the python version
-    xtric = np.random.rand(21)
-    assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:2],sym='iso')==pydis.elasticC2(c12=xtric[0],c44=xtric[1],voigt=True))
-    assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:3],sym='cubic')==pydis.elasticC2(c11=xtric[0],c12=xtric[1],c44=xtric[2],voigt=True))
-    assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:5],sym='hcp')==pydis.elasticC2(c11=xtric[0],c12=xtric[1],c13=xtric[2],c33=xtric[3],c44=xtric[4],voigt=True))
-    assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:6],sym='tetr')==pydis.elasticC2(c11=xtric[0],c12=xtric[1],c13=xtric[2],c33=xtric[3],c44=xtric[4],c66=xtric[5],voigt=True))
-    assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:6],sym='trig')==pydis.elasticC2(cij=xtric[:6],voigt=True))
-    assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:7],sym='tetr2')==pydis.elasticC2(cij=xtric[:7],voigt=True))
-    assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:9],sym='orth')==pydis.elasticC2(cij=xtric[:9],voigt=True))
-    assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric[:13],sym='mono')==pydis.elasticC2(cij=xtric[:13],voigt=True))
-    assert np.all(pydis.subroutines.elastic_constants.elasticc2(xtric,sym='tric')==pydis.elasticC2(cij=xtric,voigt=True))
+    xtric = 0.01+20*np.random.rand(56)
+    assert np.allclose(pydis.subroutines.elastic_constants.elasticc2(xtric[:2],sym='iso'),pydis.elasticC2(c12=xtric[0],c44=xtric[1],voigt=True))
+    assert np.allclose(pydis.subroutines.elastic_constants.elasticc2(xtric[:3],sym='cubic'),pydis.elasticC2(c11=xtric[0],c12=xtric[1],c44=xtric[2],voigt=True))
+    assert np.allclose(pydis.subroutines.elastic_constants.elasticc2(xtric[:5],sym='hcp'),pydis.elasticC2(c11=xtric[0],c12=xtric[1],c13=xtric[2],c33=xtric[3],c44=xtric[4],voigt=True))
+    assert np.allclose(pydis.subroutines.elastic_constants.elasticc2(xtric[:6],sym='tetr'),pydis.elasticC2(c11=xtric[0],c12=xtric[1],c13=xtric[2],c33=xtric[3],c44=xtric[4],c66=xtric[5],voigt=True))
+    assert np.allclose(pydis.subroutines.elastic_constants.elasticc2(xtric[:6],sym='trig'),pydis.elasticC2(cij=xtric[:6],voigt=True))
+    assert np.allclose(pydis.subroutines.elastic_constants.elasticc2(xtric[:7],sym='tetr2'),pydis.elasticC2(cij=xtric[:7],voigt=True))
[…]
