+++ pydislocdyn/elasticconstants.f90
@@ -109,7 +109,7 @@ module elastic_constants
       real(sel), intent(in) :: cijk(:)
       character(*), intent(in) :: sym
       real(sel), intent(out) :: vc(6,6,6)
-      real(sel) :: x, xijk(56)
+      real(sel) :: xijk(56)
       integer :: i,j,k,n,ii,jj,kk,iind(6)
       iind = (/0,15,25,31,34,35/)
       xijk = 0.d0
@@ -142,6 +142,70 @@ module elastic_constants
           xijk(19)=xijk(21); xijk(31)=xijk(21); xijk(36)=xijk(21); xijk(41)=xijk(21); xijk(44)=xijk(21) ! c155=c244=c266=c344=c355=c166
           xijk(22)=xijk(1); xijk(37)=xijk(1) ! c222=c333=c111
           xijk(34)=xijk(16); xijk(46)=xijk(16) ! c255=c366=c144
+        case ("hcp")
+          if (n/=10) then
+            print*,"error: expected length 10"
+            return
+          end if
+          ! user passes (c111,c112,c113,c123,c133,c144,c155,c222,c333,c344) in this case:
+          xijk(:3) = cijk(:3); xijk(8) = cijk(4); xijk(12) = cijk(5); xijk(16) = cijk(6); xijk(19) = cijk(7)
+          xijk(22) = cijk(8); xijk(37) = cijk(9); xijk(41) = cijk(10)
+          xijk(23) = cijk(3); xijk(27) = cijk(5); xijk(31) = cijk(7); xijk(34) = cijk(6); xijk(44) = cijk(10)
+          xijk(7) = (xijk(1)+xijk(2)-xijk(22)) !c122 = c111+c112-c222
+          xijk(21) = 0.25d0*(3.d0*xijk(22)-xijk(2)-2.d0*xijk(1)) !c166 = (3*c222-c112-2*c111)/4
+          xijk(36) = 0.25d0*(2*xijk(1)-xijk(2)-xijk(22)) !c266 = (2*c111-c112-c222)/4
+          xijk(46) = 0.5d0*(xijk(3)-xijk(8)) !c366 = (c113-c123)/2
+          xijk(51) = 0.5d0*(xijk(19)-xijk(16)) !c456 = (c155-c144)/2
+        case ("tetr")
+          if (n/=12) then
+            print*,"error: expected length 12"
+            return
+          end if
+          ! user passes (c111,c112,c113,c123,c133,c144,c155,c166,c333,c344,c366,c456) in this case:
+          xijk(:3) = cijk(:3); xijk(8) = cijk(4); xijk(12) = cijk(5); xijk(16) = cijk(6); xijk(19) = cijk(7)
+          xijk(21) = cijk(8); xijk(37) = cijk(9); xijk(41) = cijk(10); xijk(46) = cijk(11); xijk(51) = cijk(12)
+          xijk(22)=xijk(1); xijk(7) = xijk(2); xijk(23) = xijk(3); xijk(27) = xijk(12) ! c222=c111, c122=c112, c223=c113, c233=c133
+          xijk(31) = xijk(19); xijk(34) = xijk(16); xijk(36) = xijk(21); xijk(44) = xijk(41) ! c244=c155, c255=c144, c266=c166, c355=c344
+        case ("trig")
+          if (n/=14) then
[…]
