+++ pydislocdyn/dislocations.f90
@@ -1,34 +1,38 @@
 ! Author: Daniel N. Blaschke
 ! Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-! Date: Mar. 31, 2026 - Apr. 7, 2026
+! Date: Mar. 31, 2026 - Apr. 8, 2026
 module dislocations
   use parameters, only : sel, pi ! defined in subroutines.f90
+  use elastic_constants ! defined in elasticconstants.f90
   implicit none
   private
   type, public :: metalprops
     character(:), allocatable :: sym, metal ! sym defines the symmetry via keyword, metal is a name given to this instance
-    real(sel), allocatable :: cij(:) ! store linearly independent elastic constants only
-    real(sel) :: rho, C2(6,6), C3(6,6,6) ! density and elastic constants
-    real(sel) :: lat_a(3), lat_angles(3) ! lattice constants and angles
-    real(sel) :: Vc ! unit cell volume
+    real(sel), allocatable :: cij(:), cijk(:) ! store linearly independent elastic constants only
+    real(sel) :: rho=0.d0, C2(6,6)=0.d0, C3(6,6,6)=0.d0 ! density and elastic constants
+    real(sel) :: lat_a(3)=0.d0, lat_angles(3)=0.d0 ! lattice constants and angles
+    real(sel) :: Vc=0.d0 ! unit cell volume
     contains
       procedure :: update_Vc => volume_unitcell ! define as type-bound procedure
   end type metalprops
   type, extends(metalprops), public :: disloc
-    real(sel) :: b(3), n0(3) ! Burgers vector and slip plane normal
-    real(sel) :: burgers ! Burgers vector length
+    real(sel) :: b(3)=0.d0, n0(3)=0.d0 ! Burgers vector and slip plane normal
+    real(sel) :: burgers=0.d0 ! Burgers vector length
     integer :: ntheta=2, nphi=500 ! number of character angles between 0 and pi/2; resolution in polar angle phi
     real(sel), allocatable :: theta(:), phi(:), rot(:,:,:), t(:,:)
     real(sel), allocatable :: m0(:,:), M(:,:,:), N(:,:,:), Cv(:,:,:,:,:)
     contains
       procedure :: update_theta => set_character_angles
+      procedure :: update_stroh => computestroh
+      procedure :: update_rot => computerot
+      procedure :: init => init_disloc
   end type
-  public :: volume_unitcell, set_character_angles, computerot, computestroh
+  public :: volume_unitcell, set_character_angles, computerot
   contains
     subroutine volume_unitcell(mat)
       class(metalprops), intent(inout) :: mat
       select case (trim(mat%sym))
-        case ("iso","cubic")
+        case ("iso","cubic", "fcc", "bcc")
           mat%Vc = mat%lat_a(1)**3
         case ("hcp")
           mat%Vc = mat%lat_a(1)*mat%lat_a(1)*mat%lat_a(3)*sqrt(3.d0)*3.d0/2.d0
@@ -49,11 +53,15 @@ module dislocations
     subroutine set_character_angles(disl)
       class(disloc), intent(inout) :: disl
       ! todo: use select case to determine whether or not we need negative theta as well
+      if (allocated(disl%theta)) deallocate(disl%theta)
       allocate(disl%theta(disl%ntheta))
       call linspace(0.d0,pi/2.d0,disl%ntheta,disl%theta)
     end subroutine set_character_angles
     subroutine computestroh(disl)
       class(disloc), intent(inout) :: disl
[…]
