+++ pydislocdyn/subroutines.f90
@@ -2,13 +2,13 @@
 ! run 'python -m numpy.f2py -c subroutines.f90 -m subroutines' to use
 ! Author: Daniel N. Blaschke
 ! Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-! Date: July 23, 2018 - Apr. 7, 2026
+! Date: July 23, 2018 - Apr. 8, 2026
 
 module parameters
 implicit none
 integer,parameter :: sel = selected_real_kind(10)
 integer,parameter :: selsm = selected_real_kind(6)  ! some memory-heavy subroutines use lower precision in favor of speed
-integer,parameter :: version = 20260407
+integer,parameter :: version = 20260408
 real(kind=sel), parameter :: hbar = 1.0545718d-34       ! reduced Planck constant
 real(kind=sel), parameter :: kB = 1.38064852d-23        ! Boltzmann constant
 real(kind=sel), parameter :: pi = (4.d0*atan(1.d0)) ! pi
@@ -220,31 +220,6 @@ SUBROUTINE accscrew_xyintegrand(integrand,x,y,t,xpr,a,b,c,ct,abc,ca,xcomp)
   
   RETURN
 END SUBROUTINE accscrew_xyintegrand
-
-!!**********************************************************************
-
-SUBROUTINE elasticA3(C2, C3, A3)
-! Returns the tensor of elastic constants as it enters the interaction of dislocations with phonons.
-! Required inputs are the tensors of SOEC and TOEC.
-!-----------------------------------------------------------------------
-  use parameters, only : sel
-  IMPLICIT NONE
-!-----------------------------------------------------------------------
-  REAL(KIND=sel), INTENT(IN)  :: C2(3,3,3,3), C3(3,3,3,3,3,3)
-  REAL(KIND=sel), INTENT(OUT) :: A3(3,3,3,3,3,3)
-  INTEGER :: i
-  REAL(KIND=sel), DIMENSION(3,3,3,3) :: C2swap
-  
-  C2swap = reshape(C2, (/ 3, 3, 3, 3/), order = (/2,3,1,4/) )
-  A3 = C3
-  do i=1,3
-    A3(:,:,i,:,i,:) = A3(:,:,i,:,i,:) + C2
-    A3(i,:,:,:,i,:) = A3(i,:,:,:,i,:) + C2swap
-    A3(i,:,i,:,:,:) = A3(i,:,i,:,:,:) + C2
-  end do
-  
-  RETURN
-END SUBROUTINE elasticA3
 
 !!**********************************************************************
 
@@ -774,300 +749,330 @@ module phononwind_subroutines
 
 end module phononwind_subroutines
 
-!!**********************************************************************
+  !!**********************************************************************
 
-SUBROUTINE phononwind_xx(dij,A3,qBZ,ct,cl,beta,burgers,Temp,lentheta,lent,lenph,lenq1,lenph1,updatet,r0cut,debye,dragb)
-! this is a subroutine of dragcoeff_iso() in phononwind.py (TT and LL modes)
-!-----------------------------------------------------------------------
-  use parameters, only : sel, selsm, hbar, kb, pi
-  use phononwind_subroutines
-  IMPLICIT NONE
[…]
