+++ pydislocdyn/dislocations.f90
@@ -14,6 +14,7 @@ module dislocations
     real(sel) :: Vc=0.d0 ! unit cell volume
     real(sel) :: lam=0.d0, mu=0.d0 ! Lame constants (polycryst. averages)
     real(sel) :: C2norm(3,3,3,3)=0.d0 ! will be used to store unvoigt(C2)/mu
+    real(sel) :: Temp=300.d0 ! temperature associated with C2, rho, etc.
     contains
       procedure :: update_Vc => volume_unitcell ! define as type-bound procedure
   end type metalprops
@@ -31,7 +32,8 @@ module dislocations
       procedure :: init => init_disloc
       procedure :: update_uij => compute_uij
   end type
-  public :: volume_unitcell, set_character_angles, computerot
+  public :: volume_unitcell, set_character_angles, computerot, phonondrag
+  !-------------------------
   contains
     subroutine volume_unitcell(mat)
       class(metalprops), intent(inout) :: mat
@@ -82,6 +84,7 @@ module dislocations
         disl%rot(3,:,th) = disl%t(:,th)
       end do
     end subroutine computerot
+    !-------------------------
     subroutine init_disloc(disl)
       class(disloc), intent(inout) :: disl
       real(sel) :: tmp_len
@@ -127,10 +130,95 @@ module dislocations
       end if
       call unvoigt(disl%C2/disl%mu,disl%C2norm)
     end subroutine init_disloc
+    !-------------------------
     subroutine compute_uij(disl)
       class(disloc), intent(inout) :: disl
       if (allocated(disl%uij)) deallocate(disl%uij)
       allocate(disl%uij(disl%nphi,3,3,disl%ntheta))
       call computeuij(disl%beta,disl%C2norm,disl%Cv,disl%b,disl%M,disl%N,disl%phi,disl%ntheta,disl%nphi,disl%uij)
     end subroutine compute_uij
+    !-------------------------
+    subroutine phonondrag(drag,disl,beta,nphi,nq)
+      use phononwind
+      class(disloc), intent(in) :: disl
+      real(sel), intent(in) :: beta(:)
+      real(sel), intent(out), allocatable :: drag(:,:)
+      integer, optional :: nphi, nq
+      real(sel) :: uij(disl%nphi,3,3,disl%ntheta), C3norm(3,3,3,3,3,3), A3(3,3,3,3,3,3), A3rot(3,3,3,3,3,3,disl%ntheta)
+      real(sel) :: rot(3,3), uijaligned(disl%nphi,3,3,disl%ntheta), ct, cl, qBZ
+      real(sel), allocatable :: phi(:), q(:), sincos(:,:), fourieruij(:,:,:,:), dragTT(:), dragLL(:), dragTL(:), dragLT(:)
+      real(sel), allocatable :: dij(:,:,:,:)
+      integer :: lenph, lenq, th, nbeta, bt, lent, ntdyn, i, ii, j, jj, k, kk, l, ll
+      if (present(nphi)) then
+        lenph = nphi
+      else
+        lenph = 50
+      end if
+      if (present(nq)) then
+        lenq = nq
+      else
+        lenq = 50
+      end if
+      nbeta = size(beta)
+      lent = 321 ! todo: make this user-configurable
+      allocate(phi(lenph),q(lenq),sincos(disl%nphi,lenph),fourieruij(3,3,disl%ntheta,lenph),drag(disl%ntheta,nbeta))
[…]
