+++ pydislocdyn/dislocations/general.py
@@ -1,7 +1,7 @@
 # Compute various properties of a moving dislocation
 # Author: Daniel N. Blaschke
 # Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-# Date: Nov. 3, 2017 - Feb. 16, 2026
+# Date: Nov. 3, 2017 - Apr. 9, 2026
 '''This submodule contains the Dislocation class which inherits from the StrohGeometry class and the metal_props class.
    As such, it is the most complete class to compute properties of dislocations, both steady state and accelerating.
    Additionally, the Dislocation class can calculate properties like limiting velocities of dislocations. We also define
@@ -18,7 +18,8 @@ from ..crystals import metal_props, loadinputfile
 from .steadystate import StrohGeometry, elbrak1d
 
 if usefortran:
-    from ..subroutines import accscrew_xyintegrand
+    from ..subroutines import various_subroutines
+    accscrew_xyintegrand = various_subroutines.accscrew_xyintegrand
 else:
     from .numba_subroutines import accscrew_xyintegrand
 
