### 2026-04-09 e8e581c
refactor subroutines.f90 into modules
