+++ pydislocdyn/subroutines.f90
@@ -2,484 +2,450 @@
 ! run 'python -m numpy.f2py -c subroutines.f90 -m subroutines' to use
 ! Author: Daniel N. Blaschke
 ! Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-! Date: July 23, 2018 - Apr. 8, 2026
+! Date: July 23, 2018 - Apr. 9, 2026
 
 module parameters
-implicit none
-integer,parameter :: sel = selected_real_kind(10)
-integer,parameter :: selsm = selected_real_kind(6)  ! some memory-heavy subroutines use lower precision in favor of speed
-integer,parameter :: version = 20260408
-real(kind=sel), parameter :: hbar = 1.0545718d-34       ! reduced Planck constant
-real(kind=sel), parameter :: kB = 1.38064852d-23        ! Boltzmann constant
-real(kind=sel), parameter :: pi = (4.d0*atan(1.d0)) ! pi
-real(kind=sel), parameter :: pi2 = (4.d0*atan(1.d0))**2 ! pi squared
+  implicit none
+  integer,parameter :: sel = selected_real_kind(10)
+  integer,parameter :: selsm = selected_real_kind(6)  ! some memory-heavy subroutines use lower precision in favor of speed
+  integer,parameter :: version = 20260409
+  real(kind=sel), parameter :: hbar = 1.0545718d-34       ! reduced Planck constant
+  real(kind=sel), parameter :: kB = 1.38064852d-23        ! Boltzmann constant
+  real(kind=sel), parameter :: pi = (4.d0*atan(1.d0)) ! pi
+  real(kind=sel), parameter :: pi2 = (4.d0*atan(1.d0))**2 ! pi squared
 end module parameters
 
-subroutine ompinfo(nthreads)
-!$   Use omp_lib
-integer, intent(out) :: nthreads
-nthreads = 0
-!$   nthreads = omp_get_max_threads()
-!~ !$   print*, 'OpenMP: using ',nthreads,' of ',omp_get_num_procs(),' processors'
-!~ !$   print*, 'type "export OMP_NUM_THREADS=n" before running this prog. to change'
-return
-end subroutine ompinfo
+module utilities
+  implicit none
+  public :: ompinfo, elbrak, elbrak1d, cross, trapz, cumtrapz, inv, linspace
+  contains
+    subroutine ompinfo(nthreads)
+    !$   Use omp_lib
+    integer, intent(out) :: nthreads
+    nthreads = 0
+    !$   nthreads = omp_get_max_threads()
+    !~ !$   print*, 'OpenMP: using ',nthreads,' of ',omp_get_num_procs(),' processors'
+    !~ !$   print*, 'type "export OMP_NUM_THREADS=n" before running this prog. to change'
+    return
+    end subroutine ompinfo
 
-SUBROUTINE elbrak(a,b,Cmat,Ntheta,Nphi,AB)
-! Compute the bracket (A,B) := A.Cmat.B, where Cmat is a tensor of 2nd order elastic constants.
-! All three variables have an additional disloc. character dependence.
-!-----------------------------------------------------------------------
-  use parameters, only : sel
-  IMPLICIT NONE
-!-----------------------------------------------------------------------
-  INTEGER, INTENT(IN) :: Ntheta,Nphi
-  REAL(KIND=sel), DIMENSION(Nphi,3,Ntheta), INTENT(IN)  :: a, b
-  REAL(KIND=sel), DIMENSION(3,3,3,3,Ntheta), INTENT(IN)  :: Cmat
[…]
