+++ testing/test_unit.py
@@ -2,7 +2,7 @@
 # test suite for PyDislocDyn
 # Author: Daniel N. Blaschke
 # Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-# Date: Mar. 6, 2023 - Apr. 7, 2026
+# Date: Mar. 6, 2023 - Apr. 9, 2026
 '''This script implements several unit tests for PyDislocyn meant to be called by pytest.'''
 import os
 import sys
@@ -15,6 +15,8 @@ tmppydislocdyn = pathlib.Path(testpath,"temp_pydislocdyn")
 tmppydislocdyn.mkdir(exist_ok=True)
 
 import pydislocdyn as pydis
+if pydis.usefortran:
+    from pydislocdyn.subroutines import elastic_constants, utilities, various_subroutines
 import numpy as np
 from scipy.integrate import cumulative_trapezoid, trapezoid
 import sympy as sp
@@ -159,7 +161,7 @@ def test_disloc_props(metal_list=None,Ntheta=2):
             print(f"Warning: numerical accuracy of vcrit_edge for {X} may be less than 1%")
         if pydis.usefortran:
             ## check fortran implementation of stroh geometry also:
-            t,m0,M,N,Cv = pydis.subroutines.strohgeometry(Y[X].b,Y[X].n0,Y[X].theta,Y[X].phi)
+            t,m0,M,N,Cv = various_subroutines.strohgeometry(Y[X].b,Y[X].n0,Y[X].theta,Y[X].phi)
             assert np.allclose(m0,Y[X].m0.T)
             assert np.allclose(t,Y[X].t.T)
             assert np.allclose(Cv,Y[X].Cv)
@@ -181,9 +183,9 @@ def test_disloc_props(metal_list=None,Ntheta=2):
         ## check that fortran implementation of averaging schemes matches the python implementations
         for X in Y:
             if Y[X].sym in ('cubic','fcc','bcc'):
-                assert np.allclose(np.array(Y[X].compute_Lame(scheme='improved',roundto=0)[:2]),np.array(pydis.subroutines.elastic_constants.kroeneraverage(Y[X].C2)))
+                assert np.allclose(np.array(Y[X].compute_Lame(scheme='improved',roundto=0)[:2]),np.array(elastic_constants.kroeneraverage(Y[X].C2)))
             if X in ('Cu','Tibasal','Sn','Mo'): # check only one rep. for each symmetry (Mo is isotropic in contrast to Mo110 etc.)
-                assert np.allclose(np.array(Y[X].compute_Lame(scheme='hill',roundto=0)[:2]),np.array(pydis.subroutines.elastic_constants.hillaverage(Y[X].C2)))
+                assert np.allclose(np.array(Y[X].compute_Lame(scheme='hill',roundto=0)[:2]),np.array(elastic_constants.hillaverage(Y[X].C2)))
 
 def test_fortransubroutines():
     '''tests some of the fortran code, if it is available'''
@@ -192,45 +194,45 @@ def test_fortransubroutines():
         return
     # test inv()
     A = np.random.rand(9).reshape((3,3))
-    Ainv = pydis.subroutines.inv(A)
+    Ainv = utilities.inv(A)
     assert np.all(np.abs(A@Ainv-pydis.utilities.delta)<1e-9) and np.all(np.abs(np.linalg.inv(A)-Ainv)<1e-9)
     # test linspace()
[…]

[…]
