### 2026-04-10 1047b49
merge a changed subroutine from master
and update dislocations.f90 accordingly
