### 2026-04-10 c924118
changed some defaults, revert __init__.py, minor cleanup
