### 2026-04-13 6194d77
fortran lib: partial support reading input files, demo frontend
updated build systems, try fpm also on windows runner
