+++ .github/workflows/CI.yml
@@ -30,7 +30,7 @@ jobs:
       - uses: actions/checkout@v6
       - name: Setup Fortran Compiler
         uses: fortran-lang/setup-fortran@v1
-        if: runner.os == 'MacOS'
+        if: runner.os == 'MacOS' || runner.os == 'Windows'
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v6
         with:
@@ -40,14 +40,21 @@ jobs:
           python -m pip install --upgrade pip
           pip install ${{ matrix.module-versions }}
           pip install .
+      - name: Set up Fortran Package Manager
+        uses: fortran-lang/setup-fpm@v7
+        with:
+            github-token: ${{ secrets.GITHUB_TOKEN }}
       - name: Run tests
         run: |
           make
           ./runtests.x
+          fpm test
           yes 2>/dev/null | pytest -s --durations=0 testing/
         if: runner.os != 'Windows'
       - name: Run tests windows
-        run: echo 'N' | pytest -s --durations=0 testing/
+        run:  |
+          fpm test
+          echo 'N' | pytest -s --durations=0 testing/
         if: runner.os == 'Windows'
       - name: Upload log files as artifacts
         uses: actions/upload-artifact@v7
