### 2026-04-13 6c0b84a
compatibility update for older fpm 0.12
