+++ .github/workflows/CI.yml
@@ -40,20 +40,18 @@ jobs:
           python -m pip install --upgrade pip
           pip install ${{ matrix.module-versions }}
           pip install .
-      - name: Set up Fortran Package Manager
-        uses: fortran-lang/setup-fpm@v9
-        with:
-            github-token: ${{ secrets.GITHUB_TOKEN }}
       - name: Run tests
         run: |
           make
           ./runtests.x
+          pip install fpm
           fpm build --verbose
           fpm test
           yes 2>/dev/null | pytest -s --durations=0 testing/
         if: runner.os != 'Windows'
       - name: Run tests windows
         run:  |
+          pip install fpm
           fpm build --verbose
           fpm test
           echo 'N' | pytest -s --durations=0 testing/
