### 2026-04-17 a3f139c
add do concurrent in some places
