### 2026-04-17 7f406e0
build system updates: use hatchling instead of setuptools
TODO: update docs if we stick with it
