+++ .github/workflows/CI.yml
@@ -30,7 +30,7 @@ jobs:
       - uses: actions/checkout@v6
       - name: Setup Fortran Compiler
         uses: fortran-lang/setup-fortran@v1
-        if: runner.os == 'MacOS' || runner.os == 'Windows'
+        if: runner.os == 'MacOS'
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v6
         with:
@@ -44,17 +44,10 @@ jobs:
         run: |
           make
           ./runtests.x
-          pip install fpm
-          fpm build --verbose
-          fpm test
           yes 2>/dev/null | pytest -s --durations=0 testing/
         if: runner.os != 'Windows'
       - name: Run tests windows
-        run:  |
-          pip install fpm
-          fpm build --verbose
-          fpm test
-          echo 'N' | pytest -s --durations=0 testing/
+        run: echo 'N' | pytest -s --durations=0 testing/
         if: runner.os == 'Windows'
       - name: Upload log files as artifacts
         uses: actions/upload-artifact@v7
