### 2026-04-18 bcef90b
promote inv, cross to operators
