+++ pydislocdyn/subroutines.f90
@@ -2,13 +2,13 @@
 ! run 'python -m numpy.f2py -c subroutines.f90 -m subroutines' to use
 ! Author: Daniel N. Blaschke
 ! Copyright (c) 2018, Triad National Security, LLC. All rights reserved.
-! Date: July 23, 2018 - Apr. 17, 2026
+! Date: July 23, 2018 - Apr. 18, 2026
 
 module parameters
   implicit none
   integer,parameter :: sel = selected_real_kind(10)
   integer,parameter :: selsm = selected_real_kind(6)  ! some memory-heavy subroutines use lower precision in favor of speed
-  integer,parameter :: version = 20260417
+  integer,parameter :: version = 20260418
   real(kind=sel), parameter :: hbar = 1.0545718d-34       ! reduced Planck constant
   real(kind=sel), parameter :: kB = 1.38064852d-23        ! Boltzmann constant
   real(kind=sel), parameter :: pi = (4.d0*atan(1.d0)) ! pi
@@ -17,7 +17,13 @@ end module parameters
 
 module utilities
   implicit none
-  public :: ompinfo, elbrak, elbrak1d, cross, trapz, cumtrapz, inv, linspace
+  interface operator(.cross.)
+    module procedure cross
+  end interface
+  interface operator(.inv.)
+    module procedure inv
+  end interface
+  public :: ompinfo, elbrak, elbrak1d, cross, operator(.cross.), trapz, cumtrapz, inv, operator(.inv.), linspace
   contains
     subroutine ompinfo(nthreads)
     !$   Use omp_lib
@@ -58,10 +64,9 @@ module utilities
       end do
       !$OMP END PARALLEL DO
       
-      RETURN
     END SUBROUTINE elbrak
 
-    SUBROUTINE elbrak1d(a,b,Cmat,Nphi,AB)
+    pure SUBROUTINE elbrak1d(a,b,Cmat,Nphi,AB)
     ! Compute the bracket (A,B) := A.Cmat.B, where Cmat is a tensor of 2nd order elastic constants.
     !-----------------------------------------------------------------------
       use parameters, only : sel
@@ -85,24 +90,23 @@ module utilities
         end do
       end do
       
-      RETURN
     END SUBROUTINE elbrak1d
 
     !!**********************************************************************
 
-    pure subroutine cross(x,y,z)
+    pure function cross(x,y) result(z)
       use parameters, only : sel
       implicit none
       real(sel), dimension(3), intent(in)  :: x, y
-      real(sel), dimension(3), intent(out)  :: z
+      real(sel), dimension(3)  :: z
       z(1) = x(2)*y(3) - x(3)*y(2)
       z(2) = x(3)*y(1) - x(1)*y(3)
       z(3) = x(1)*y(2) - x(2)*y(1)
-    end subroutine cross
+    end function cross
 
     !!**********************************************************************
 
-    SUBROUTINE trapz(f,x,n,intf)
+    pure SUBROUTINE trapz(f,x,n,intf)
     ! integrate using the trapezoidal rule
     !-----------------------------------------------------------------------
       use parameters, only : sel
@@ -113,13 +117,11 @@ module utilities
       REAL(KIND=sel), INTENT(OUT) :: intf
      
       intf = sum(0.5d0*(f(2:n)+f(1:n-1))*(x(2:n)-x(1:n-1)))
-      
-      RETURN
     END SUBROUTINE trapz
 
     !!**********************************************************************
 
-    SUBROUTINE cumtrapz(f,x,n,intf)
+    pure SUBROUTINE cumtrapz(f,x,n,intf)
     ! cumulatively integrate using the trapezoidal rule
     !-----------------------------------------------------------------------
       use parameters, only : sel
[…]
