### 2026-04-22 7f9d6aa
some code optimization
