+++ pydislocdyn/readinputfiles.f90
@@ -80,17 +80,17 @@ module readinputfiles
       ! start postprocessing / initializing depending on symmetry
       select case (trim(disl%sym))
         case ("iso")
-          disl%cij = (/c12,c44/)
-          disl%cijk = (/c123,c144,c456/)
+          disl%cij = [c12,c44]
+          disl%cijk = [c123,c144,c456]
         case ("cubic", "fcc", "bcc")
-          disl%cij = (/c11,c12,c44/)
-          disl%cijk = (/c111,c112,c123,c144,c166,c456/)
+          disl%cij = [c11,c12,c44]
+          disl%cijk = [c111,c112,c123,c144,c166,c456]
         case ("hcp")
-          disl%cij = (/c11,c12,c13,c33,c44/)
-          disl%cijk = (/c111,c112,c113,c123,c133,c144,c155,c222,c333,c344/)
+          disl%cij = [c11,c12,c13,c33,c44]
+          disl%cijk = [c111,c112,c113,c123,c133,c144,c155,c222,c333,c344]
         case ("tetr")
-          disl%cij = (/c11,c12,c13,c33,c44,c66/)
-          disl%cijk = (/c111,c112,c113,c123,c133,c144,c155,c166,c333,c344,c366,c456/)
+          disl%cij = [c11,c12,c13,c33,c44,c66]
+          disl%cijk = [c111,c112,c113,c123,c133,c144,c155,c166,c333,c344,c366,c456]
         case default
           print*,"Error: keyword sym must be one of 'iso', 'cubic', 'hcp', 'tetr', 'trig', 'tetr2', 'orth', 'mono', 'tric'."
           return
