### 2026-04-23 9c53247
workaround for lfortran and nicer error messages
